Fractional Wishart Helper Function

An expansion of R's 'stats' random wishart matrix generation. This package allows the user to generate singular, Uhlig and Harald (1994) <doi:10.1214/aos/1176325375>, and pseudo wishart, Diaz-Garcia, et al.(1997) <doi:10.1006/jmva.1997.1689>, matrices. In addition the user can generate wishart matrices with fractional degrees of freedom, Adhikari (2008) <doi:10.1061/(ASCE)0733-9399(2008)134:12(1029)>, commonly used in volatility modeling. Users can also use this package to create random covariance matrices.

Generate n random matrices, distributed according to the Wishart distribution with parameters Sigma and df, W_p(Sigma, df).

FractionalWishart(df, Sigma, covariance = FALSE)

Arguments

df
numeric parameter, “degrees of freedom”.

Sigma
positive definite (\(p\times p\)) “scale” matrix, the matrix parameter of the distribution.

covariance
logical on whether a covariance matrix should be generated

Value

A numeric array of dimension p * p * n, where each array is a positive semidefinite matrix, a realization of the Wishart distribution W_p(Sigma, df)

Details

If X_1, ..., X_m is a sample of m independent multivariate Gaussians with mean vector 0, and covariance matrix Sigma, the distribution of M = X'X is W_p(Sigma, m).

Examples

FractionalWishart(22.5, diag(1, 20))

#>              [,1]         [,2]        [,3]         [,4]       [,5]        [,6]
#>  [1,]  92.8176833 -15.38075606  -2.0818683 -14.06145908  9.2536127   0.3372104
#>  [2,] -15.3807561 119.88962890   3.4081550 -14.11264852  1.3729232   3.9242227
#>  [3,]  -2.0818683   3.40815495  55.7911901  18.59532621 -0.6245865 -11.0041558
#>  [4,] -14.0614591 -14.11264852  18.5953262 110.43220174 -0.8341097   1.0358559
#>  [5,]   9.2536127   1.37292322  -0.6245865  -0.83410972 66.6206980   5.1067622
#>  [6,]   0.3372104   3.92422273 -11.0041558   1.03585588  5.1067622  54.3840669
#>  [7,]  14.5562389  -3.61133012  -3.7451233  -0.45066729  9.5658109   6.8635483
#>  [8,]   6.2689198  -0.04956875   8.8602985   3.66855142 -0.1160564 -15.4116364
#>  [9,] -11.2027320   5.71539322  -6.0015089   9.10053664 -8.0538895  -3.5743580
#> [10,] -10.5029170   1.16549136  -7.8702176  -0.06645995 -0.1166403   3.9443563
#> [11,]  -0.5931502  -3.45355343   9.9423774  -6.02455248  3.1126842  -6.0373545
#> [12,]  12.4470716  -8.33043142 -12.8232322  -2.34618361  2.8054089  -1.3471233
#> [13,]  14.6329628   2.52066972   0.8102732  -4.33497560 13.3094099   9.7659071
#> [14,]  -1.1125407   0.05801636   8.0104830  -2.23243299 -0.7689147  -7.6313991
#> [15,]  -4.6334997  -2.19074427   1.3991770  -2.11473474 -5.6613035   9.4230606
#> [16,]  -0.7229132  -3.03889489   3.1798680  -1.50547861  1.1964046  -9.0818014
#> [17,]  -2.8788398  -7.06335769   2.6060890   0.80692910 -2.4896777  -5.2300819
#> [18,]  -0.1861320   1.59124301   1.2930910  -1.03941008 -3.5367941  -0.5969805
#> [19,]   1.0119421  -1.68838813   0.3366329   0.07482180  2.0704637  -1.0931675
#> [20,]   1.9705754   2.14805132   1.0592459  -0.74305801  2.3543283  -0.1226214
#>              [,7]         [,8]        [,9]        [,10]      [,11]        [,12]
#>  [1,] 14.55623895   6.26891985 -11.2027320 -10.50291701 -0.5931502  12.44707162
#>  [2,] -3.61133012  -0.04956875   5.7153932   1.16549136 -3.4535534  -8.33043142
#>  [3,] -3.74512331   8.86029850  -6.0015089  -7.87021759  9.9423774 -12.82323224
#>  [4,] -0.45066729   3.66855142   9.1005366  -0.06645995 -6.0245525  -2.34618361
#>  [5,]  9.56581092  -0.11605642  -8.0538895  -0.11664034  3.1126842   2.80540893
#>  [6,]  6.86354829 -15.41163642  -3.5743580   3.94435628 -6.0373545  -1.34712331
#>  [7,] 58.08642528   0.92426744  -6.0363378   3.85645624  3.5179126  -3.18997220
#>  [8,]  0.92426744  66.83304173  -1.7542687 -12.24228151  7.1231585   2.76793146
#>  [9,] -6.03633776  -1.75426873  48.9385980  18.50441878  3.8802827  -0.47741243
#> [10,]  3.85645624 -12.24228151  18.5044188  40.74418252 -0.9247914  13.26299186
#> [11,]  3.51791262   7.12315846   3.8802827  -0.92479141 76.8561518   7.37528311
#> [12,] -3.18997220   2.76793146  -0.4774124  13.26299186  7.3752831  37.60317945
#> [13,] -0.03365421  -3.24912218  -2.8277689  -3.99855295 -5.8616763  -5.51705022
#> [14,] -3.80875218  -2.34708879 -12.6148780  -2.91244275 -1.2916958  -1.18794483
#> [15,] -6.10463305 -11.93416810   7.4834421   2.36435535 11.0329831   1.55211448
#> [16,]  4.31636781  -2.86180147  -8.0810737  -4.98906250  2.3597911  -5.39669499
#> [17,]  1.34323632   1.35502161  -3.2154872   0.09087772  2.4205272   1.51761299
#> [18,] -6.55891871  -6.52086836   5.4870895  -2.08495993  1.8551595  -1.56995685
#> [19,]  2.06011676   1.13750800  -1.9901541   0.32508136 -0.5786437  -0.04354951
#> [20,] -0.47097194  -4.23406962   0.1133270  -1.90100999 -1.0957381  -3.57122979
#>             [,13]        [,14]       [,15]      [,16]       [,17]      [,18]
#>  [1,] 14.63296282  -1.11254068  -4.6334997 -0.7229132 -2.87883978 -0.1861320
#>  [2,]  2.52066972   0.05801636  -2.1907443 -3.0388949 -7.06335769  1.5912430
#>  [3,]  0.81027323   8.01048296   1.3991770  3.1798680  2.60608901  1.2930910
#>  [4,] -4.33497560  -2.23243299  -2.1147347 -1.5054786  0.80692910 -1.0394101
#>  [5,] 13.30940995  -0.76891470  -5.6613035  1.1964046 -2.48967765 -3.5367941
#>  [6,]  9.76590708  -7.63139906   9.4230606 -9.0818014 -5.23008192 -0.5969805
#>  [7,] -0.03365421  -3.80875218  -6.1046331  4.3163678  1.34323632 -6.5589187
#>  [8,] -3.24912218  -2.34708879 -11.9341681 -2.8618015  1.35502161 -6.5208684
#>  [9,] -2.82776893 -12.61487796   7.4834421 -8.0810737 -3.21548717  5.4870895
#> [10,] -3.99855295  -2.91244275   2.3643554 -4.9890625  0.09087772 -2.0849599
#> [11,] -5.86167629  -1.29169580  11.0329831  2.3597911  2.42052722  1.8551595
#> [12,] -5.51705022  -1.18794483   1.5521145 -5.3966950  1.51761299 -1.5699569
#> [13,] 26.06507295  -2.12545940   2.1617678 -3.7363536 -8.12905313 -0.3968790
#> [14,] -2.12545940  38.70250348  -0.7391165  4.0747211  3.65865663 -1.1916423
#> [15,]  2.16176785  -0.73911652  30.3434564 -4.1256516 -1.93545713 11.3542953
#> [16,] -3.73635363   4.07472113  -4.1256516 16.8781067  5.10295185 -2.4501605
#> [17,] -8.12905313   3.65865663  -1.9354571  5.1029519  7.12890776 -1.9763003
#> [18,] -0.39687903  -1.19164230  11.3542953 -2.4501605 -1.97630027 10.4713297
#> [19,]  0.24194711   1.08499861  -4.1075240  1.8197336  1.20256539 -2.8756390
#> [20,]  4.58407413  -1.99599825  -0.5894864  1.3608293 -2.19866958  0.9381959
#>             [,19]      [,20]
#>  [1,]  1.01194207  1.9705754
#>  [2,] -1.68838813  2.1480513
#>  [3,]  0.33663289  1.0592459
#>  [4,]  0.07482180 -0.7430580
#>  [5,]  2.07046373  2.3543283
#>  [6,] -1.09316755 -0.1226214
#>  [7,]  2.06011676 -0.4709719
#>  [8,]  1.13750800 -4.2340696
#>  [9,] -1.99015411  0.1133270
#> [10,]  0.32508136 -1.9010100
#> [11,] -0.57864373 -1.0957381
#> [12,] -0.04354951 -3.5712298
#> [13,]  0.24194711  4.5840741
#> [14,]  1.08499861 -1.9959983
#> [15,] -4.10752403 -0.5894864
#> [16,]  1.81973357  1.3608293
#> [17,]  1.20256539 -2.1986696
#> [18,] -2.87563895  0.9381959
#> [19,]  1.21629668  0.1299735
#> [20,]  0.12997352  2.9947573
#> attr(,"df")
#> ~22.5
#> <environment: namespace:rWishart>