Psuedo Wishart Helper Function

An expansion of R's 'stats' random wishart matrix generation. This package allows the user to generate singular, Uhlig and Harald (1994) <doi:10.1214/aos/1176325375>, and pseudo wishart, Diaz-Garcia, et al.(1997) <doi:10.1006/jmva.1997.1689>, matrices. In addition the user can generate wishart matrices with fractional degrees of freedom, Adhikari (2008) <doi:10.1061/(ASCE)0733-9399(2008)134:12(1029)>, commonly used in volatility modeling. Users can also use this package to create random covariance matrices.

Generate n random matrices, distributed according to the Wishart distribution with parameters Sigma and df, W_p(Sigma, df).

PsuedoWishart(df, Sigma, covariance = FALSE)

Arguments

df
numeric parameter, “degrees of freedom”.

Sigma
positive definite (\(p\times p\)) “scale” matrix, the matrix parameter of the distribution.

covariance
logical on whether a covariance matrix should be generated

Value

A numeric array of dimension p * p * n, where each array is a positive semidefinite matrix, a realization of the Wishart distribution W_p(Sigma, df)

Details

If X_1, ..., X_m is a sample of m independent multivariate Gaussians with mean vector 0, and covariance matrix Sigma, the distribution of M = X'X is W_p(Sigma, m).

Examples

PsuedoWishart(5, diag(1, 20))

#>             [,1]        [,2]        [,3]         [,4]       [,5]       [,6]
#>  [1,]  7.0441018  0.74165046  1.53863047 -2.999346421  2.0617169  1.3000963
#>  [2,]  0.7416505  8.81565849  1.58621527  0.766215183 -2.1429650  4.1013035
#>  [3,]  1.5386305  1.58621527  3.87993035  2.053532540 -2.2339528  1.6596019
#>  [4,] -2.9993464  0.76621518  2.05353254  4.924135228 -5.5996967 -1.8688782
#>  [5,]  2.0617169 -2.14296500 -2.23395276 -5.599696670  8.0474240  2.5110834
#>  [6,]  1.3000963  4.10130350  1.65960191 -1.868878177  2.5110834  5.8727456
#>  [7,]  4.6414925  6.12959766  0.75429915 -1.852111374 -0.1399763  2.7696355
#>  [8,]  3.4919013  6.32741055 -0.36598481 -2.203051789  1.3214040  2.3513454
#>  [9,] -0.7053406 -1.95476288 -3.25254971 -2.735615270  3.0660618 -0.4636047
#> [10,]  3.6719317 -0.94528482  0.46530042 -1.561493014  0.6380805 -0.1486713
#> [11,]  3.6394645 -1.68359723  0.30322169 -2.702918928  2.7785170  0.9858143
#> [12,] -1.3944711 -0.87720376  2.89553326  2.242182918 -2.9425048  2.1136056
#> [13,]  4.3523863 -0.63953073 -0.72734105 -2.254904543  1.8732374 -1.6724086
#> [14,] -7.7993811  0.58335645 -2.10277319  2.423316826 -0.5806589 -0.3335008
#> [15,]  3.7201823  2.24132607  0.51834941 -0.007163651 -1.8698223 -1.3883970
#> [16,] -1.3434422  3.71381634 -0.06521107  2.370552185 -3.8894228 -1.0367683
#> [17,] -1.0755798 -0.33781778 -2.91758760 -1.074779338  0.9969461 -1.7871455
#> [18,]  0.8515847 -5.10556190 -0.77046276 -2.671087390  3.7485738  0.4328304
#> [19,] -2.9171426 -0.09936575 -0.06218490 -1.528103622  4.4145915  4.3735159
#> [20,]  1.2997441  1.57915929  1.95565136  1.673234000 -1.3464396 -0.7637962
#>              [,7]       [,8]       [,9]      [,10]      [,11]      [,12]
#>  [1,]  4.64149246  3.4919013 -0.7053406  3.6719317  3.6394645 -1.3944711
#>  [2,]  6.12959766  6.3274105 -1.9547629 -0.9452848 -1.6835972 -0.8772038
#>  [3,]  0.75429915 -0.3659848 -3.2525497  0.4653004  0.3032217  2.8955333
#>  [4,] -1.85211137 -2.2030518 -2.7356153 -1.5614930 -2.7029189  2.2421829
#>  [5,] -0.13997626  1.3214040  3.0660618  0.6380805  2.7785170 -2.9425048
#>  [6,]  2.76963551  2.3513454 -0.4636047 -0.1486713  0.9858143  2.1136056
#>  [7,]  7.27134530  6.6802247 -0.7164905  1.8102555  1.0523962 -2.1154396
#>  [8,]  6.68022474  7.8743812 -0.2426309  0.5557716  0.1363619 -5.2127817
#>  [9,] -0.71649049 -0.2426309  3.2492679  0.1079108  0.6627392 -1.3661420
#> [10,]  1.81025551  0.5557716  0.1079108  2.4488446  2.2657316  0.2827832
#> [11,]  1.05239617  0.1363619  0.6627392  2.2657316  2.8694585  0.3998083
#> [12,] -2.11543962 -5.2127817 -1.3661420  0.2827832  0.3998083  8.7025189
#> [13,]  2.61213832  3.2235683  0.3539411  2.1482143  1.8438814 -4.6797805
#> [14,] -4.35647614 -1.9012280  0.9708558 -4.7693494 -4.2963381 -0.5780732
#> [15,]  4.16798063  3.8456701 -1.1142000  1.8566755  0.5490504 -2.8768487
#> [16,]  2.07552370  2.3985173 -1.0580324 -1.0221793 -2.4005848 -1.3439872
#> [17,]  0.09385611  1.3433876  2.0365440 -0.5342678 -0.7509916 -3.4426219
#> [18,] -2.85163919 -3.8855194  2.0902436  1.4059356  2.7171595  2.6721127
#> [19,] -2.73853334 -1.6847763  1.0339865 -2.2673732 -0.1003787  2.0462338
#> [20,]  1.02961779  2.1206239 -2.6218004 -0.2575568 -0.7103376 -2.7211257
#>            [,13]      [,14]        [,15]       [,16]       [,17]      [,18]
#>  [1,]  4.3523863 -7.7993811  3.720182289 -1.34344218 -1.07557981  0.8515847
#>  [2,] -0.6395307  0.5833565  2.241326065  3.71381634 -0.33781778 -5.1055619
#>  [3,] -0.7273411 -2.1027732  0.518349414 -0.06521107 -2.91758760 -0.7704628
#>  [4,] -2.2549045  2.4233168 -0.007163651  2.37055219 -1.07477934 -2.6710874
#>  [5,]  1.8732374 -0.5806589 -1.869822335 -3.88942279  0.99694606  3.7485738
#>  [6,] -1.6724086 -0.3335008 -1.388396966 -1.03676833 -1.78714548  0.4328304
#>  [7,]  2.6121383 -4.3564761  4.167980631  2.07552370  0.09385611 -2.8516392
#>  [8,]  3.2235683 -1.9012280  3.845670067  2.39851732  1.34338759 -3.8855194
#>  [9,]  0.3539411  0.9708558 -1.114199978 -1.05803242  2.03654398  2.0902436
#> [10,]  2.1482143 -4.7693494  1.856675541 -1.02217930 -0.53426777  1.4059356
#> [11,]  1.8438814 -4.2963381  0.549050450 -2.40058480 -0.75099156  2.7171595
#> [12,] -4.6797805 -0.5780732 -2.876848710 -1.34398718 -3.44262191  2.6721127
#> [13,]  4.8275450 -4.3325606  3.543920069 -0.19841561  1.17237985 -0.3803668
#> [14,] -4.3325606  9.8076232 -4.213457357  1.59530588  1.73681395 -1.7840314
#> [15,]  3.5439201 -4.2134574  4.574659731  2.04593464  0.53109576 -2.6145719
#> [16,] -0.1984156  1.5953059  2.045934640  3.68556655  0.97086111 -4.0961928
#> [17,]  1.1723798  1.7368140  0.531095759  0.97086111  2.57912080 -0.7098641
#> [18,] -0.3803668 -1.7840314 -2.614571936 -4.09619278 -0.70986407  5.4161116
#> [19,] -3.5769293  4.6112472 -5.195138826 -2.72056058 -0.93554937  2.4721576
#> [20,]  1.8993157 -0.6423582  2.157557223  1.23887503 -0.52999375 -2.9222603
#>             [,19]      [,20]
#>  [1,] -2.91714257  1.2997441
#>  [2,] -0.09936575  1.5791593
#>  [3,] -0.06218490  1.9556514
#>  [4,] -1.52810362  1.6732340
#>  [5,]  4.41459154 -1.3464396
#>  [6,]  4.37351595 -0.7637962
#>  [7,] -2.73853334  1.0296178
#>  [8,] -1.68477629  2.1206239
#>  [9,]  1.03398647 -2.6218004
#> [10,] -2.26737317 -0.2575568
#> [11,] -0.10037868 -0.7103376
#> [12,]  2.04623376 -2.7211257
#> [13,] -3.57692932  1.8993157
#> [14,]  4.61124722 -0.6423582
#> [15,] -5.19513883  2.1575572
#> [16,] -2.72056058  1.2388750
#> [17,] -0.93554937 -0.5299937
#> [18,]  2.47215760 -2.9222603
#> [19,]  7.93622484 -1.7153823
#> [20,] -1.71538231  4.0258284
#> attr(,"df")
#> ~5
#> <environment: namespace:rWishart>