Singular Wishart Helper Function

An expansion of R's 'stats' random wishart matrix generation. This package allows the user to generate singular, Uhlig and Harald (1994) <doi:10.1214/aos/1176325375>, and pseudo wishart, Diaz-Garcia, et al.(1997) <doi:10.1006/jmva.1997.1689>, matrices. In addition the user can generate wishart matrices with fractional degrees of freedom, Adhikari (2008) <doi:10.1061/(ASCE)0733-9399(2008)134:12(1029)>, commonly used in volatility modeling. Users can also use this package to create random covariance matrices.

Generate n random matrices, distributed according to the Wishart distribution with parameters Sigma and df, W_p(Sigma, df).

SingularWishart(df, Sigma, covariance = FALSE)

Arguments

df
numeric parameter, “degrees of freedom”.

Sigma
positive definite (\(p\times p\)) “scale” matrix, the matrix parameter of the distribution.

covariance
logical on whether a covariance matrix should be generated

Value

A numeric array of dimension p * p * n, where each array is a positive semidefinite matrix, a realization of the Wishart distribution W_p(Sigma, df)

Details

If X_1, ..., X_m is a sample of m independent multivariate Gaussians with mean vector 0, and covariance matrix Sigma, the distribution of M = X'X is W_p(Sigma, m).

Examples

SingularWishart(5, diag(1, 20))

#>              [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#>  [1,]  2.02224937 -2.80095962  1.09781092 -0.02187379 -1.30296894 -1.11599642
#>  [2,] -2.80095962 14.22438907 -2.40856515  0.01553288  7.13525060 -1.88986771
#>  [3,]  1.09781092 -2.40856515  1.76626268 -1.15281951 -1.29844626  1.33227670
#>  [4,] -0.02187379  0.01553288 -1.15281951  5.70707243 -1.03480766 -6.57108341
#>  [5,] -1.30296894  7.13525060 -1.29844626 -1.03480766  3.92809951  0.01279426
#>  [6,] -1.11599642 -1.88986771  1.33227670 -6.57108341  0.01279426  9.48971776
#>  [7,]  0.61612173  4.97332249  0.18027091  1.95294300  1.92324447 -4.25139789
#>  [8,]  2.67445393  0.88187148  1.65090978 -1.07643624  0.50227309 -1.55029462
#>  [9,]  0.30324369 -4.74286033  0.42146801 -2.23201963 -1.97247177  3.74258439
#> [10,]  0.87067786 -2.92294918  0.81766248 -1.12025551 -1.28094469  1.39374596
#> [11,]  1.45413211 -2.80395276  0.33660084  1.88834799 -1.63302849 -2.79106970
#> [12,] -0.45907977  3.40577711 -0.69882242  1.78674783  1.18941829 -2.58549636
#> [13,]  0.06332816  1.84846287  0.46102011  0.68067878  0.70842968 -1.29039542
#> [14,]  1.47794586  1.98816545  1.84694226  0.19961148  0.73648586 -1.99987788
#> [15,] -1.39977209  0.56536396 -0.47216409 -1.97455941  0.51510007  3.51975576
#> [16,]  3.45049301 -8.08787792  4.17614266 -1.34929348 -4.41906115  1.39508428
#> [17,]  0.83835887 -3.50442613  0.14856794 -0.27059906 -1.46337836  0.40262232
#> [18,] -0.51200121  6.85625047 -0.81912634  0.96914838  3.04961894 -2.69723686
#> [19,] -1.18938514 -4.03129849 -0.30083236  2.19106238 -2.75491169  0.19560610
#> [20,]  1.14547737  3.14711369  0.01386146  2.67336062  1.11104264 -5.20616736
#>             [,7]       [,8]       [,9]      [,10]        [,11]        [,12]
#>  [1,]  0.6161217  2.6744539  0.3032437  0.8706779  1.454132113 -0.459079771
#>  [2,]  4.9733225  0.8818715 -4.7428603 -2.9229492 -2.803952756  3.405777105
#>  [3,]  0.1802709  1.6509098  0.4214680  0.8176625  0.336600837 -0.698822421
#>  [4,]  1.9529430 -1.0764362 -2.2320196 -1.1202555  1.888347992  1.786747830
#>  [5,]  1.9232445  0.5022731 -1.9724718 -1.2809447 -1.633028486  1.189418292
#>  [6,] -4.2513979 -1.5502946  3.7425844  1.3937460 -2.791069697 -2.585496361
#>  [7,]  4.8373446  3.6912039 -3.2631952 -1.0183523  0.529523891  2.437223525
#>  [8,]  3.6912039  7.7781259  0.5265339  1.0898067  1.133094257  1.939400728
#>  [9,] -3.2631952  0.5265339  5.0121893  1.8180243 -0.137726412 -0.409417296
#> [10,] -1.0183523  1.0898067  1.8180243  1.0172649  0.381589620 -0.575922331
#> [11,]  0.5295239  1.1330943 -0.1377264  0.3815896  1.763539781  0.004121119
#> [12,]  2.4372235  1.9394007 -0.4094173 -0.5759223  0.004121119  2.779862907
#> [13,]  1.6664032 -0.5894575 -3.3059502 -0.9244592  0.036740400 -0.639177710
#> [14,]  3.6270947  2.0608929 -4.8071430 -0.7448830  0.634563717 -0.893702385
#> [15,] -1.8377896 -0.8018131  2.7165221  0.4042677 -1.552033989  0.482533478
#> [16,]  0.3035093  4.5115790  1.6797330  2.3491164  2.004734536 -1.447964040
#> [17,] -1.6614983 -0.2816953  1.4185742  0.7764167  0.786462506 -1.123972072
#> [18,]  4.1624068  4.0401164 -1.0137379 -0.8616322 -0.577203395  3.883156468
#> [19,] -2.4087228 -3.4751889  2.0544183  0.2408549  0.240231549  0.231452073
#> [20,]  4.1686129  2.9102064 -3.4486530 -0.9270351  1.311615712  1.597203205
#>             [,13]      [,14]      [,15]       [,16]      [,17]      [,18]
#>  [1,]  0.06332816  1.4779459 -1.3997721  3.45049301  0.8383589 -0.5120012
#>  [2,]  1.84846287  1.9881654  0.5653640 -8.08787792 -3.5044261  6.8562505
#>  [3,]  0.46102011  1.8469423 -0.4721641  4.17614266  0.1485679 -0.8191263
#>  [4,]  0.68067878  0.1996115 -1.9745594 -1.34929348 -0.2705991  0.9691484
#>  [5,]  0.70842968  0.7364859  0.5151001 -4.41906115 -1.4633784  3.0496189
#>  [6,] -1.29039542 -1.9998779  3.5197558  1.39508428  0.4026223 -2.6972369
#>  [7,]  1.66640317  3.6270947 -1.8377896  0.30350930 -1.6614983  4.1624068
#>  [8,] -0.58945750  2.0608929 -0.8018131  4.51157896 -0.2816953  4.0401164
#>  [9,] -3.30595025 -4.8071430  2.7165221  1.67973299  1.4185742 -1.0137379
#> [10,] -0.92445919 -0.7448830  0.4042677  2.34911639  0.7764167 -0.8616322
#> [11,]  0.03674040  0.6345637 -1.5520340  2.00473454  0.7864625 -0.5772034
#> [12,] -0.63917771 -0.8937024  0.4825335 -1.44796404 -1.1239721  3.8831565
#> [13,]  2.68353583  4.1743898 -2.0411533  0.44079520 -0.7754864 -0.5535471
#> [14,]  4.17438978  7.8564686 -3.7389503  3.60354153 -1.0552848 -0.2162420
#> [15,] -2.04115331 -3.7389503  2.9626624 -1.65482990 -0.1426071  0.7303443
#> [16,]  0.44079520  3.6035415 -1.6548299 10.89465923  1.0714454 -2.2496903
#> [17,] -0.77548638 -1.0552848 -0.1426071  1.07144535  1.2662207 -1.9377613
#> [18,] -0.55354706 -0.2162420  0.7303443 -2.24969033 -1.9377613  6.3481687
#> [19,] -1.41998004 -3.6043104  1.4092147  0.06651244  0.4314793 -1.6455101
#> [20,]  1.86814029  3.7940738 -2.7238164  0.42588951 -0.9112741  2.6428864
#>             [,19]       [,20]
#>  [1,] -1.18938514  1.14547737
#>  [2,] -4.03129849  3.14711369
#>  [3,] -0.30083236  0.01386146
#>  [4,]  2.19106238  2.67336062
#>  [5,] -2.75491169  1.11104264
#>  [6,]  0.19560610 -5.20616736
#>  [7,] -2.40872284  4.16861293
#>  [8,] -3.47518894  2.91020637
#>  [9,]  2.05441832 -3.44865300
#> [10,]  0.24085492 -0.92703508
#> [11,]  0.24023155  1.31161571
#> [12,]  0.23145207  1.59720320
#> [13,] -1.41998004  1.86814029
#> [14,] -3.60431035  3.79407379
#> [15,]  1.40921466 -2.72381643
#> [16,]  0.06651244  0.42588951
#> [17,]  0.43147934 -0.91127405
#> [18,] -1.64551008  2.64288640
#> [19,]  5.16103270 -2.30739078
#> [20,] -2.30739078  4.28760379
#> attr(,"df")
#> ~5
#> <environment: namespace:rWishart>