Random Fractional Wishart Matrix

Random Fractional Wishart Matrix

rFractionalWishart(n, df, Sigma, covariance = FALSE, simplify = "array")

Arguments

n

integer: the number of replications.

df

degrees of freedom

Sigma

Covariance matrix

covariance

should covariance matrix be generated (X / df)

simplify

logical or character string; should the result be simplified to a vector, matrix or higher dimensional array if possible? For sapply it must be named and not abbreviated. The default value, TRUE, returns a vector or matrix if appropriate, whereas if simplify = "array" the result may be an array of “rank” (\(=\)length(dim(.))) one higher than the result of FUN(X[[i]]).

Value

list or array of random wishart matrices

Examples

rFractionalWishart(2, 22.5, diag(1, 20))

#> , , 1
#> 
#>              [,1]        [,2]         [,3]        [,4]        [,5]        [,6]
#>  [1,] 122.7848383  -4.2289983   4.01996660  -0.4800401   9.9175227  -5.9322659
#>  [2,]  -4.2289983 156.2031906   4.55293683  10.6084320   8.6489884   3.2090014
#>  [3,]   4.0199666   4.5529368  81.20340957 -10.0163144   0.7435996   9.8996805
#>  [4,]  -0.4800401  10.6084320 -10.01631436  61.3984493   1.4065996 -16.0487689
#>  [5,]   9.9175227   8.6489884   0.74359964   1.4065996 149.3900951  -9.3712072
#>  [6,]  -5.9322659   3.2090014   9.89968049 -16.0487689  -9.3712072  79.8335700
#>  [7,]  -0.9106847  -3.8065526  -4.02233820   1.4705015   4.3282485  -8.0612954
#>  [8,]  19.9592184  -2.1919655  11.00258809 -14.7639083  -3.7719057  12.7754369
#>  [9,]  -7.8680427  -6.2358937 -12.11548159   1.0371042   9.6682595  -0.8570272
#> [10,]   3.7970362  -2.2633576  -0.61352298  -0.1392381   5.3842624   0.3810409
#> [11,]  -4.5347509   8.9607606  17.81981893  -4.5371793  -3.5158753  20.3042702
#> [12,]  -4.7211442  -7.2955034  -3.88076553  -0.9466947  -7.6452998   1.6980478
#> [13,]   8.6419188  -6.6182961   1.81756452   2.0788530  -0.6217965   2.0233627
#> [14,]   2.2502836  -1.3933876  -1.68912989   7.8937305  -8.7446219  -0.7073153
#> [15,]  -0.9292733  -3.8564852  -8.37697130   1.6412433   1.4135387  -6.6684631
#> [16,]   4.9432619  -4.5065697  -0.94867728   8.4752555  -3.1412419   6.3794129
#> [17,]   0.8899564   2.7321149   4.00243650  -2.8729292   2.7279615  -3.0390572
#> [18,]  -6.4650890   1.3092236   0.06735198  -0.5109184  -3.2347560  -2.5695040
#> [19,]   0.1345999   1.8484496   1.64671185   0.3350216   2.6694786  -1.3025525
#> [20,]   0.8201546  -0.7831971  -0.33606137  -2.1075676   0.5636374   1.3025120
#>              [,7]        [,8]        [,9]      [,10]        [,11]        [,12]
#>  [1,]  -0.9106847  19.9592184  -7.8680427  3.7970362  -4.53475094 -4.721144230
#>  [2,]  -3.8065526  -2.1919655  -6.2358937 -2.2633576   8.96076056 -7.295503403
#>  [3,]  -4.0223382  11.0025881 -12.1154816 -0.6135230  17.81981893 -3.880765535
#>  [4,]   1.4705015 -14.7639083   1.0371042 -0.1392381  -4.53717929 -0.946694743
#>  [5,]   4.3282485  -3.7719057   9.6682595  5.3842624  -3.51587530 -7.645299751
#>  [6,]  -8.0612954  12.7754369  -0.8570272  0.3810409  20.30427022  1.698047794
#>  [7,]  46.4591695  -1.4688511  -7.2720282  0.8087122 -17.73302848 12.054699453
#>  [8,]  -1.4688511  66.7809182  -3.7189314  6.4228182   9.82783451 -4.461262355
#>  [9,]  -7.2720282  -3.7189314  91.3345351 -6.5316239  -1.16067488 -5.718443522
#> [10,]   0.8087122   6.4228182  -6.5316239 48.9217240  -3.87044046 -0.235210045
#> [11,] -17.7330285   9.8278345  -1.1606749 -3.8704405  88.10456685 -7.323735380
#> [12,]  12.0546995  -4.4612624  -5.7184435 -0.2352100  -7.32373538 31.073754331
#> [13,]   9.1408694   7.9471137  -5.1726920  4.0181827  -0.20573995 -0.593658210
#> [14,]   0.4131710  -2.9762307   3.5381492 -1.1759223   2.73866662  4.820058434
#> [15,]   3.2179838  -1.2928363   3.6059600 -2.7846835  -5.88812607 11.677204730
#> [16,] -10.3836977   0.9049102  -2.8553549  0.4302864   3.86341090  0.002355005
#> [17,]   0.3440909   0.2552833   1.9735707 -0.9401908   0.02055647 -5.601857508
#> [18,]  -2.3055906  -4.6362667   1.7213804 -0.8603615   4.38724915 -3.620703918
#> [19,]  -1.9834392  -1.6896072   0.2524023  0.6747538   0.33167179 -2.401267443
#> [20,]   1.7276689   1.6967018  -1.5343641  0.4655914  -0.27666278  1.057785659
#>            [,13]      [,14]       [,15]         [,16]       [,17]       [,18]
#>  [1,]  8.6419188  2.2502836 -0.92927333   4.943261891  0.88995637 -6.46508898
#>  [2,] -6.6182961 -1.3933876 -3.85648520  -4.506569666  2.73211485  1.30922356
#>  [3,]  1.8175645 -1.6891299 -8.37697130  -0.948677284  4.00243650  0.06735198
#>  [4,]  2.0788530  7.8937305  1.64124328   8.475255459 -2.87292918 -0.51091842
#>  [5,] -0.6217965 -8.7446219  1.41353873  -3.141241938  2.72796154 -3.23475603
#>  [6,]  2.0233627 -0.7073153 -6.66846307   6.379412908 -3.03905720 -2.56950404
#>  [7,]  9.1408694  0.4131710  3.21798380 -10.383697728  0.34409095 -2.30559060
#>  [8,]  7.9471137 -2.9762307 -1.29283629   0.904910233  0.25528325 -4.63626665
#>  [9,] -5.1726920  3.5381492  3.60596003  -2.855354895  1.97357067  1.72138044
#> [10,]  4.0181827 -1.1759223 -2.78468350   0.430286433 -0.94019076 -0.86036147
#> [11,] -0.2057400  2.7386666 -5.88812607   3.863410903  0.02055647  4.38724915
#> [12,] -0.5936582  4.8200584 11.67720473   0.002355005 -5.60185751 -3.62070392
#> [13,] 45.9115966  6.6413014 -6.94365154  -5.298255890 -1.87654171  1.24701769
#> [14,]  6.6413014 23.9511110  0.90399174  -0.340788354  0.06731590  0.74878022
#> [15,] -6.9436515  0.9039917 32.63099053   2.048686802 -0.21790919  0.50728844
#> [16,] -5.2982559 -0.3407884  2.04868680  23.998295028 -3.36714230 -2.07336952
#> [17,] -1.8765417  0.0673159 -0.21790919  -3.367142299  5.70080271  1.29065459
#> [18,]  1.2470177  0.7487802  0.50728844  -2.073369520  1.29065459 10.61740751
#> [19,] -0.3511105 -0.5533774 -0.21340245   2.154124767  1.96356929  1.93788513
#> [20,]  1.8308969 -0.3668359 -0.01444828  -1.780833256 -1.52889904 -1.49965753
#>            [,19]       [,20]
#>  [1,]  0.1345999  0.82015456
#>  [2,]  1.8484496 -0.78319711
#>  [3,]  1.6467118 -0.33606137
#>  [4,]  0.3350216 -2.10756759
#>  [5,]  2.6694786  0.56363736
#>  [6,] -1.3025525  1.30251201
#>  [7,] -1.9834392  1.72766888
#>  [8,] -1.6896072  1.69670181
#>  [9,]  0.2524023 -1.53436409
#> [10,]  0.6747538  0.46559137
#> [11,]  0.3316718 -0.27666278
#> [12,] -2.4012674  1.05778566
#> [13,] -0.3511105  1.83089688
#> [14,] -0.5533774 -0.36683595
#> [15,] -0.2134025 -0.01444828
#> [16,]  2.1541248 -1.78083326
#> [17,]  1.9635693 -1.52889904
#> [18,]  1.9378851 -1.49965753
#> [19,]  2.9281401 -1.77151765
#> [20,] -1.7715177  2.06716749
#> 
#> , , 2
#> 
#>              [,1]        [,2]         [,3]        [,4]         [,5]       [,6]
#>  [1,] 162.3022165   0.9656253  -6.89102411  17.8239528 -16.22722588  0.1991864
#>  [2,]   0.9656253 229.3401164 -17.52122533  -6.0594619  18.88312039  6.2139414
#>  [3,]  -6.8910241 -17.5212253 110.21689721   4.7623376   3.14939321  3.4389929
#>  [4,]  17.8239528  -6.0594619   4.76233760 105.4637025 -27.15821366  0.7495857
#>  [5,] -16.2272259  18.8831204   3.14939321 -27.1582137 171.87763866 -8.7238955
#>  [6,]   0.1991864   6.2139414   3.43899294   0.7495857  -8.72389549 92.5930193
#>  [7,] -13.2174768 -11.7308007   8.15094441  -5.7320329 -15.70452469 14.5834136
#>  [8,]  11.3284274  -7.6311694 -14.38824869   1.5855711  -3.55809131 12.8284357
#>  [9,]  -8.1050927  -5.9216575  -0.02094372  -1.8933554   8.87625468 10.3333730
#> [10,]  -1.3778660   3.9127339   3.34591003  -1.2464419   5.20093463  4.2693672
#> [11,]  -0.3785763 -26.7139786  -8.55721521   4.3968950 -14.48741238  3.9999925
#> [12,]  -4.3213861  -0.4543918  -3.78856591  -0.1164793  -7.20360774  1.5689851
#> [13,]  -5.3057058  -6.6400533   4.65617417  -4.8986236   6.91513979  2.2154928
#> [14,]  -5.1009284  -8.2228789   4.58505461   1.4926204   5.75695106 -2.5979449
#> [15,]  -0.1975127  -7.6608627   6.12649640  -3.3206340  -3.32797272  3.6159779
#> [16,]   1.9502996   6.6578162  -6.58346265   0.9467385  -1.58004431 -1.8682446
#> [17,]   0.5965661   4.2688305  -0.44381537   6.9600316  -0.70854493 -2.9385707
#> [18,]   0.4561872   4.2628900  -1.16607737   1.1264846  -2.09100052 -1.6041318
#> [19,]  -0.4970712  -0.9173403   0.92060289   0.6748196   0.08163609 -1.8035646
#> [20,]   1.0068564  -0.3112977  -0.41119568   0.4226691  -0.07786461  1.3359949
#>              [,7]        [,8]        [,9]        [,10]       [,11]      [,12]
#>  [1,] -13.2174768  11.3284274 -8.10509268  -1.37786602  -0.3785763 -4.3213861
#>  [2,] -11.7308007  -7.6311694 -5.92165749   3.91273392 -26.7139786 -0.4543918
#>  [3,]   8.1509444 -14.3882487 -0.02094372   3.34591003  -8.5572152 -3.7885659
#>  [4,]  -5.7320329   1.5855711 -1.89335544  -1.24644195   4.3968950 -0.1164793
#>  [5,] -15.7045247  -3.5580913  8.87625468   5.20093463 -14.4874124 -7.2036077
#>  [6,]  14.5834136  12.8284357 10.33337298   4.26936724   3.9999925  1.5689851
#>  [7,] 139.0621543   8.4233029 -6.54991615 -16.45116682   6.5326217  1.0367333
#>  [8,]   8.4233029 123.0601885  7.15180889 -13.40246603   7.5284271 -1.5802821
#>  [9,]  -6.5499162   7.1518089 67.44351375   8.17136224   1.8155141  3.2828380
#> [10,] -16.4511668 -13.4024660  8.17136224  56.95195471  -8.0937526  0.5020996
#> [11,]   6.5326217   7.5284271  1.81551406  -8.09375257  81.7724339  3.7243961
#> [12,]   1.0367333  -1.5802821  3.28283798   0.50209957   3.7243961 18.7035923
#> [13,]  -3.5202332  -1.9111043  3.64142543   0.88267749  -2.5800334  0.1920143
#> [14,]  -6.5418380  -3.5617947  9.17299965   2.11262103   1.3875957  4.8232880
#> [15,]   3.9320171  -0.9281500  4.69553355   1.78721594   3.5594752 -2.7111985
#> [16,]  -0.7048979   2.9863400 -3.06846133   3.34522993   0.4692581 -6.2711979
#> [17,]  -0.2544139  -6.5944699  0.79763387   1.46733438  -2.8806482  2.8421170
#> [18,]  -2.5775392  -0.9085472 -2.58158869  -0.90021868   2.0108283  0.6604450
#> [19,]  -0.9056747   1.1246277  1.14360526   0.60023100  -0.9529858 -1.1189502
#> [20,]   0.8926852  -0.6694144 -1.02251447  -0.08134128   0.6164815  0.3120150
#>            [,13]      [,14]       [,15]       [,16]       [,17]       [,18]
#>  [1,] -5.3057058 -5.1009284 -0.19751273  1.95029962  0.59656610  0.45618718
#>  [2,] -6.6400533 -8.2228789 -7.66086268  6.65781621  4.26883049  4.26288995
#>  [3,]  4.6561742  4.5850546  6.12649640 -6.58346265 -0.44381537 -1.16607737
#>  [4,] -4.8986236  1.4926204 -3.32063398  0.94673848  6.96003155  1.12648462
#>  [5,]  6.9151398  5.7569511 -3.32797272 -1.58004431 -0.70854493 -2.09100052
#>  [6,]  2.2154928 -2.5979449  3.61597793 -1.86824455 -2.93857071 -1.60413177
#>  [7,] -3.5202332 -6.5418380  3.93201713 -0.70489794 -0.25441393 -2.57753920
#>  [8,] -1.9111043 -3.5617947 -0.92815001  2.98634003 -6.59446991 -0.90854717
#>  [9,]  3.6414254  9.1729996  4.69553355 -3.06846133  0.79763387 -2.58158869
#> [10,]  0.8826775  2.1126210  1.78721594  3.34522993  1.46733438 -0.90021868
#> [11,] -2.5800334  1.3875957  3.55947525  0.46925807 -2.88064820  2.01082829
#> [12,]  0.1920143  4.8232880 -2.71119852 -6.27119785  2.84211702  0.66044503
#> [13,] 16.8630799  5.3581355 -4.39410038 -5.71901240 -3.56579912 -2.06502785
#> [14,]  5.3581355 18.1476116  0.47621550 -6.60535957  2.47638419 -0.30247842
#> [15,] -4.3941004  0.4762155 22.35636065 -0.07815937 -3.69001727 -0.56675037
#> [16,] -5.7190124 -6.6053596 -0.07815937 10.82026723 -2.16276726  0.87568445
#> [17,] -3.5657991  2.4763842 -3.69001727 -2.16276726 11.99541845  0.45590822
#> [18,] -2.0650279 -0.3024784 -0.56675037  0.87568445  0.45590822  2.87511516
#> [19,] -2.4180201  1.5844204  2.44886383  0.83434815  0.01135213  0.19589191
#> [20,]  0.7538972 -0.8876658 -0.69907132 -0.24187043  0.69692616 -0.06891228
#>             [,19]       [,20]
#>  [1,] -0.49707116  1.00685644
#>  [2,] -0.91734031 -0.31129769
#>  [3,]  0.92060289 -0.41119568
#>  [4,]  0.67481958  0.42266910
#>  [5,]  0.08163609 -0.07786461
#>  [6,] -1.80356460  1.33599491
#>  [7,] -0.90567467  0.89268517
#>  [8,]  1.12462773 -0.66941442
#>  [9,]  1.14360526 -1.02251447
#> [10,]  0.60023100 -0.08134128
#> [11,] -0.95298576  0.61648152
#> [12,] -1.11895024  0.31201496
#> [13,] -2.41802012  0.75389719
#> [14,]  1.58442038 -0.88766580
#> [15,]  2.44886383 -0.69907132
#> [16,]  0.83434815 -0.24187043
#> [17,]  0.01135213  0.69692616
#> [18,]  0.19589191 -0.06891228
#> [19,]  2.52055232 -1.35741417
#> [20,] -1.35741417  1.06257213
#>