Random Nonsingular Wishart Matrix

Random Nonsingular Wishart Matrix

rNonsingularWishart(n, df, Sigma, covariance = FALSE, simplify = "array")

Arguments

n
integer: the number of replications.

df
degrees of freedom

Sigma
Covariance matrix

covariance
should covariance matrix be generated (X / df)

simplify
logical or character string; should the result be simplified to a vector, matrix or higher dimensional array if possible? For sapply it must be named and not abbreviated. The default value, TRUE, returns a vector or matrix if appropriate, whereas if simplify = "array" the result may be an array of “rank” (\(=\)length(dim(.))) one higher than the result of FUN(X[[i]]).

Value

list or array of random wishart matrices

Examples

rNonsingularWishart(2, 20, diag(1, 5))

#> , , 1
#> 
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 17.2726135 -0.6760883 -3.8263078 -7.4819404  1.7835508
#> [2,] -0.6760883 10.2930600  4.0076631 -3.9248875  0.3214259
#> [3,] -3.8263078  4.0076631 14.8155221 -0.7714779 -4.2744961
#> [4,] -7.4819404 -3.9248875 -0.7714779 22.5781949 -2.9993555
#> [5,]  1.7835508  0.3214259 -4.2744961 -2.9993555 11.1977287
#> 
#> , , 2
#> 
#>             [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,]  26.0422020  0.1964692 -9.5295586  9.2623856 -10.942862
#> [2,]   0.1964692 22.2972116 -4.6410894 -4.3543525   2.206220
#> [3,]  -9.5295586 -4.6410894 19.4089824  0.8515076   1.803796
#> [4,]   9.2623856 -4.3543525  0.8515076 25.7108632  -2.241045
#> [5,] -10.9428618  2.2062202  1.8037958 -2.2410450  23.929472
#>