Random Psuedo Wishart Matrix

Random Psuedo Wishart Matrix

rPsuedoWishart(n, df, Sigma, covariance = FALSE, simplify = "array")

Arguments

n
integer: the number of replications.

df
degrees of freedom

Sigma
Covariance matrix

covariance
should covariance matrix be generated (X / df)

simplify
logical or character string; should the result be simplified to a vector, matrix or higher dimensional array if possible? For sapply it must be named and not abbreviated. The default value, TRUE, returns a vector or matrix if appropriate, whereas if simplify = "array" the result may be an array of “rank” (\(=\)length(dim(.))) one higher than the result of FUN(X[[i]]).

Value

list or array of random wishart matrices

Examples

rPsuedoWishart(2, 5, diag(1, 20))

#> , , 1
#> 
#>             [,1]        [,2]        [,3]        [,4]        [,5]       [,6]
#>  [1,]  4.3265930   1.6960812  5.19507849  1.82041339 -0.79681736 -0.9518790
#>  [2,]  1.6960812   7.4840946  1.82713634  1.86934454  2.92560726 -2.5891709
#>  [3,]  5.1950785   1.8271363  9.69772173  1.59296214 -0.11109324  0.6834580
#>  [4,]  1.8204134   1.8693445  1.59296214  3.36633101 -0.07965254 -0.4200543
#>  [5,] -0.7968174   2.9256073 -0.11109324 -0.07965254  7.95828906  3.7285069
#>  [6,] -0.9518790  -2.5891709  0.68345802 -0.42005428  3.72850687  5.0203070
#>  [7,]  6.0456821   3.8982569  4.91041046  2.94370219 -2.27938924 -3.9163892
#>  [8,] -1.2130648   0.3969254 -6.17597184  0.45988882 -1.19273488 -2.8188216
#>  [9,] -0.2906798  -3.0776373 -0.30751963  1.17512398 -1.94657556  1.5326356
#> [10,] -0.5637865   1.8821795 -1.89266172 -1.17813682  0.97524994 -1.4849870
#> [11,] -0.7474205   3.6037093  0.01536023 -0.61830899  2.22802597 -1.5333606
#> [12,] -0.3752060  -1.7707222  3.12927398 -1.15562940 -0.49029811  1.7703642
#> [13,] -0.8309346  -5.4222555 -0.83028259 -2.97043460 -3.72278701  0.5762855
#> [14,]  3.5444267   5.4295075  3.54221903  1.20629104  0.04100610 -3.7915942
#> [15,] -1.4364936  -0.4564691 -1.26439514 -0.31391697  5.60612820  4.2401325
#> [16,] -0.4876383  -3.0706977 -4.11184093  1.68427469 -6.21069092 -3.4215311
#> [17,] -4.8589408 -10.0921453 -3.44123716 -2.44650106 -1.36971379  5.8396932
#> [18,] -3.1335241  -4.9134958 -2.04050371 -2.78094431 -7.06536864 -2.1757306
#> [19,]  1.7671345  -0.1632650  0.72328835 -1.37049026 -2.37560578 -2.8507704
#> [20,]  1.2361716   1.1742869  2.66239362  0.86503247 -2.49528065 -1.8494520
#>              [,7]       [,8]         [,9]       [,10]       [,11]      [,12]
#>  [1,]   6.0456821 -1.2130648 -0.290679773 -0.56378646 -0.74742049 -0.3752060
#>  [2,]   3.8982569  0.3969254 -3.077637326  1.88217953  3.60370930 -1.7707222
#>  [3,]   4.9104105 -6.1759718 -0.307519631 -1.89266172  0.01536023  3.1292740
#>  [4,]   2.9437022  0.4598888  1.175123976 -1.17813682 -0.61830899 -1.1556294
#>  [5,]  -2.2793892 -1.1927349 -1.946575557  0.97524994  2.22802597 -0.4902981
#>  [6,]  -3.9163892 -2.8188216  1.532635576 -1.48498696 -1.53336060  1.7703642
#>  [7,]  10.6254305  1.7245852 -1.033920885  0.57849258 -0.91338273 -3.0785284
#>  [8,]   1.7245852  6.9017466 -0.286859613  1.93967801 -0.43094813 -4.8074497
#>  [9,]  -1.0339209 -0.2868596  3.712559133 -2.00772150 -4.09537093  0.4439979
#> [10,]   0.5784926  1.9396780 -2.007721501  2.05789407  1.45025785 -1.5443820
#> [11,]  -0.9133827 -0.4309481 -4.095370932  1.45025785  6.34499569  0.5878295
#> [12,]  -3.0785284 -4.8074497  0.443997894 -1.54438200  0.58782951  4.1667362
#> [13,]  -1.7347849 -0.4665867  1.853112144 -0.36144825 -3.66234536  1.4632731
#> [14,]   6.5235988  0.5294079 -3.596418607  1.63784244  3.53431410 -1.4996394
#> [15,]  -3.4207097 -0.6856628 -0.004100882 -0.07198878  0.08250263 -0.1253054
#> [16,]   1.6135923  5.2675281  1.771921906 -0.84155302 -0.86680286 -2.3706096
#> [17,] -10.4498834 -2.6080195  4.463218936 -3.44333752 -3.05259402  4.3903305
#> [18,]  -4.9368546 -1.0931552  1.003428238 -1.00412458  0.77129519  3.7932722
#> [19,]   3.6398235  1.5784467 -2.527400344  1.35761610  1.80853092 -0.8653127
#> [20,]   1.6395240 -1.5847165 -0.042633819 -0.66693540  0.80281106  1.3305568
#>            [,13]      [,14]        [,15]      [,16]      [,17]      [,18]
#>  [1,] -0.8309346  3.5444267 -1.436493634 -0.4876383  -4.858941 -3.1335241
#>  [2,] -5.4222555  5.4295075 -0.456469136 -3.0706977 -10.092145 -4.9134958
#>  [3,] -0.8302826  3.5422190 -1.264395141 -4.1118409  -3.441237 -2.0405037
#>  [4,] -2.9704346  1.2062910 -0.313916971  1.6842747  -2.446501 -2.7809443
#>  [5,] -3.7227870  0.0410061  5.606128196 -6.2106909  -1.369714 -7.0653686
#>  [6,]  0.5762855 -3.7915942  4.240132468 -3.4215311   5.839693 -2.1757306
#>  [7,] -1.7347849  6.5235988 -3.420709739  1.6135923 -10.449883 -4.9368546
#>  [8,] -0.4665867  0.5294079 -0.685662847  5.2675281  -2.608020 -1.0931552
#>  [9,]  1.8531121 -3.5964186 -0.004100882  1.7719219   4.463219  1.0034282
#> [10,] -0.3614483  1.6378424 -0.071988784 -0.8415530  -3.443338 -1.0041246
#> [11,] -3.6623454  3.5343141  0.082502627 -0.8668029  -3.052594  0.7712952
#> [12,]  1.4632731 -1.4996394 -0.125305376 -2.3706096   4.390330  3.7932722
#> [13,]  6.3913214 -3.3309928 -1.342152556  0.8267414   5.479525  4.7606272
#> [14,] -3.3309928  6.8802925 -2.406176494 -0.4752832  -9.721333 -2.8165681
#> [15,] -1.3421526 -2.4061765  5.076967000 -3.5243080   2.906854 -4.4614419
#> [16,]  0.8267414 -0.4752832 -3.524307980 10.4954825   2.507987  4.9121087
#> [17,]  5.4795250 -9.7213331  2.906853647  2.5079868  17.702170  7.7084789
#> [18,]  4.7606272 -2.8165681 -4.461441947  4.9121087   7.708479 12.6199887
#> [19,]  1.3167130  3.5335782 -2.429346112  2.3786613  -2.952184  1.5696664
#> [20,] -0.4772142  1.7415249 -2.694563492  0.7413673  -1.702681  2.4245445
#>            [,19]       [,20]
#>  [1,]  1.7671345  1.23617156
#>  [2,] -0.1632650  1.17428692
#>  [3,]  0.7232883  2.66239362
#>  [4,] -1.3704903  0.86503247
#>  [5,] -2.3756058 -2.49528065
#>  [6,] -2.8507704 -1.84945197
#>  [7,]  3.6398235  1.63952398
#>  [8,]  1.5784467 -1.58471649
#>  [9,] -2.5274003 -0.04263382
#> [10,]  1.3576161 -0.66693540
#> [11,]  1.8085309  0.80281106
#> [12,] -0.8653127  1.33055675
#> [13,]  1.3167130 -0.47721418
#> [14,]  3.5335782  1.74152492
#> [15,] -2.4293461 -2.69456349
#> [16,]  2.3786613  0.74136728
#> [17,] -2.9521839 -1.70268111
#> [18,]  1.5696664  2.42454448
#> [19,]  5.3139202  0.43359105
#> [20,]  0.4335910  2.35320900
#> 
#> , , 2
#> 
#>              [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#>  [1,] 10.84436185  0.29631560  1.19592097  4.42284747 -3.23951983 -2.97898003
#>  [2,]  0.29631560  3.69835555 -1.55535691 -0.78636146 -1.61051089 -1.13583301
#>  [3,]  1.19592097 -1.55535691  3.83598993  4.59796525 -0.03347534 -0.08504449
#>  [4,]  4.42284747 -0.78636146  4.59796525  8.14444770 -2.04208282  0.12222254
#>  [5,] -3.23951983 -1.61051089 -0.03347534 -2.04208282  5.80631975  3.15347466
#>  [6,] -2.97898003 -1.13583301 -0.08504449  0.12222254  3.15347466  6.17234588
#>  [7,] -0.02642313 -3.44022627  3.82599473  3.75907905  0.16748956 -2.56270746
#>  [8,]  4.26498155 -2.59261874  0.76851389  0.12120941  1.74514841  1.30024873
#>  [9,]  2.70478523  0.77348639 -0.54138953  1.54695574 -1.34697098 -0.91910463
#> [10,] -3.59976705 -1.81655971  1.29664244 -0.47654355  4.24808892  3.39146022
#> [11,] -2.04570521 -1.54602941  0.10336642 -0.02076276 -0.47706459  1.80934928
#> [12,] -2.08038170 -0.71289780  1.07226682 -0.58814652  0.91839478 -2.79106941
#> [13,] -4.33002601  0.68999039 -0.28096050 -1.09758057 -1.06699394 -1.00053719
#> [14,]  1.76942059  2.12049292  0.96485326  2.85395153 -3.07242472  0.49027674
#> [15,]  3.53581205 -1.10370653  0.02301026  0.68101482  3.30517242  1.28238748
#> [16,] -0.01915713  0.08965853 -0.65154634 -1.11574791 -0.80989988 -1.83802113
#> [17,]  0.20990532 -1.86288477  0.77675579  1.57934801  0.96601604  2.02226279
#> [18,]  1.10080268  0.35337082 -2.32055960 -0.81813888 -2.88691365  0.81354099
#> [19,]  1.01671822  0.67040207  1.11220839  0.75410466 -1.73412602 -2.13692729
#> [20,] -2.98667753 -3.72926228  3.25193840  2.95406722  4.25422470  6.26335769
#>              [,7]        [,8]        [,9]      [,10]       [,11]      [,12]
#>  [1,] -0.02642313  4.26498155  2.70478523 -3.5997671 -2.04570521 -2.0803817
#>  [2,] -3.44022627 -2.59261874  0.77348639 -1.8165597 -1.54602941 -0.7128978
#>  [3,]  3.82599473  0.76851389 -0.54138953  1.2966424  0.10336642  1.0722668
#>  [4,]  3.75907905  0.12120941  1.54695574 -0.4765436 -0.02076276 -0.5881465
#>  [5,]  0.16748956  1.74514841 -1.34697098  4.2480889 -0.47706459  0.9183948
#>  [6,] -2.56270746  1.30024873 -0.91910463  3.3914602  1.80934928 -2.7910694
#>  [7,]  8.58516155 -0.75612083  0.40007912  0.4850959  0.42826124  4.1162252
#>  [8,] -0.75612083  6.82166944 -1.25539355  1.6651602  0.01681829 -1.6331708
#>  [9,]  0.40007912 -1.25539355  2.68376447 -2.3252033 -0.63805860 -0.6013785
#> [10,]  0.48509592  1.66516019 -2.32520335  4.3876214  0.55344266  0.5366454
#> [11,]  0.42826124  0.01681829 -0.63805860  0.5534427  2.31095139 -0.8089292
#> [12,]  4.11622519 -1.63317078 -0.60137855  0.5366454 -0.80892921  3.8694968
#> [13,]  1.25236908 -3.96349904 -0.48719391 -0.3793215  0.84518372  1.7752993
#> [14,] -2.59482940 -0.62571033 -0.09914847 -1.1233571  0.17012791 -2.2938420
#> [15,]  0.08180799  3.31126735  1.38197471  1.1020616 -1.61510094 -0.6597262
#> [16,]  0.80827741 -0.72224121  0.29106140 -1.1182426 -0.16869703  1.0242599
#> [17,]  1.38556523  0.94584547  0.60140756  0.8508902  1.10075929 -0.7421606
#> [18,] -2.53051925 -0.42275279  1.54244731 -2.6750382  1.79518228 -2.8662627
#> [19,]  0.41713090  0.14870971 -1.00340875 -0.5753936 -0.52245809  0.8092106
#> [20,]  1.61983119  2.96030867 -2.22866534  5.5125137  2.30670024 -1.1310367
#>             [,13]        [,14]       [,15]       [,16]      [,17]       [,18]
#>  [1,] -4.33002601  1.769420591  3.53581205 -0.01915713  0.2099053  1.10080268
#>  [2,]  0.68999039  2.120492919 -1.10370653  0.08965853 -1.8628848  0.35337082
#>  [3,] -0.28096050  0.964853260  0.02301026 -0.65154634  0.7767558 -2.32055960
#>  [4,] -1.09758057  2.853951532  0.68101482 -1.11574791  1.5793480 -0.81813888
#>  [5,] -1.06699394 -3.072424720  3.30517242 -0.80989988  0.9660160 -2.88691365
#>  [6,] -1.00053719  0.490276739  1.28238748 -1.83802113  2.0222628  0.81354099
#>  [7,]  1.25236908 -2.594829404  0.08180799  0.80827741  1.3855652 -2.53051925
#>  [8,] -3.96349904 -0.625710332  3.31126735 -0.72224121  0.9458455 -0.42275279
#>  [9,] -0.48719391 -0.099148467  1.38197471  0.29106140  0.6014076  1.54244731
#> [10,] -0.37932152 -1.123357131  1.10206159 -1.11824261  0.8508902 -2.67503816
#> [11,]  0.84518372  0.170127909 -1.61510094 -0.16869703  1.1007593  1.79518228
#> [12,]  1.77529932 -2.293842020 -0.65972622  1.02425992 -0.7421606 -2.86626268
#> [13,]  3.30138549 -0.025684130 -3.46307473  0.72728606 -0.7721136  0.05190417
#> [14,] -0.02568413  4.570304972 -2.21550642 -0.87250899 -0.7527615  1.02294433
#> [15,] -3.46307473 -2.215506419  5.62216842 -0.62532381  1.4121933 -1.11948363
#> [16,]  0.72728606 -0.872508994 -0.62532381  0.83415683 -0.5157747  0.28482477
#> [17,] -0.77211359 -0.752761474  1.41219333 -0.51577471  1.8829580  0.74312279
#> [18,]  0.05190417  1.022944326 -1.11948363  0.28482477  0.7431228  4.91898983
#> [19,]  0.44833471  1.527667914 -1.81001601  0.26248226 -1.3678634 -1.11231217
#> [20,] -1.23055232 -0.003497209  1.50419759 -2.22228264  2.9172524 -1.71851056
#>            [,19]        [,20]
#>  [1,]  1.0167182 -2.986677530
#>  [2,]  0.6704021 -3.729262281
#>  [3,]  1.1122084  3.251938396
#>  [4,]  0.7541047  2.954067218
#>  [5,] -1.7341260  4.254224702
#>  [6,] -2.1369273  6.263357686
#>  [7,]  0.4171309  1.619831185
#>  [8,]  0.1487097  2.960308667
#>  [9,] -1.0034088 -2.228665341
#> [10,] -0.5753936  5.512513729
#> [11,] -0.5224581  2.306700244
#> [12,]  0.8092106 -1.131036685
#> [13,]  0.4483347 -1.230552322
#> [14,]  1.5276679 -0.003497209
#> [15,] -1.8100160  1.504197589
#> [16,]  0.2624823 -2.222282640
#> [17,] -1.3678634  2.917252433
#> [18,] -1.1123122 -1.718510561
#> [19,]  2.1188437 -1.291731463
#> [20,] -1.2917315 10.001012977
#>