Random Singular Wishart Matrix

Random Singular Wishart Matrix

rSingularWishart(n, df, Sigma, covariance = FALSE, simplify = "array")

Arguments

n

integer: the number of replications.

df

degrees of freedom

Sigma

Covariance matrix

covariance

should covariance matrix be generated (X / df)

simplify

logical or character string; should the result be simplified to a vector, matrix or higher dimensional array if possible? For sapply it must be named and not abbreviated. The default value, TRUE, returns a vector or matrix if appropriate, whereas if simplify = "array" the result may be an array of “rank” (\(=\)length(dim(.))) one higher than the result of FUN(X[[i]]).

Value

list or array of random wishart matrices

Examples

rSingularWishart(2, 5, diag(1, 20))

#> , , 1
#> 
#>              [,1]        [,2]       [,3]       [,4]        [,5]        [,6]
#>  [1,]  2.02463174 -0.53038693 -1.3159829  1.4077501  1.05593124 -0.94221918
#>  [2,] -0.53038693  0.73494526  0.4763100  0.7124516 -1.03150032  1.04789253
#>  [3,] -1.31598286  0.47631000 22.6594948  7.2234181 -1.56549168  1.48556527
#>  [4,]  1.40775013  0.71245162  7.2234181  8.2928491  1.20296213  1.34966149
#>  [5,]  1.05593124 -1.03150032 -1.5654917  1.2029621  5.40322105 -0.67240881
#>  [6,] -0.94221918  1.04789253  1.4855653  1.3496615 -0.67240881  1.74685361
#>  [7,] -1.10951371 -0.19866350 -2.5772689 -3.3521367  0.48320219 -0.08115338
#>  [8,]  0.62226314  0.78005270 -5.5952136 -0.8600820 -2.69483214  0.35802890
#>  [9,]  0.47593743  0.37353174 -5.6759716  0.6794684  3.24225454  1.08780011
#> [10,] -3.00802262  0.55652553 -0.8115165 -4.3547601 -3.56293977  0.42433039
#> [11,]  2.42025320 -1.48817879 -0.9728255  1.1141970  4.18929182 -1.79667533
#> [12,]  0.31930243  0.92589568 -4.1797729  0.9547393  0.07778905  1.32476980
#> [13,]  2.23308398 -1.24530650 -3.4684265  0.8612286  2.51444185 -2.00116177
#> [14,]  2.97010403  0.10929929  5.7044306  5.7993972  1.24425063  0.42335563
#> [15,]  0.69615297 -0.77546783 -3.4265740 -2.5327763  0.70520795 -1.24538736
#> [16,]  0.39549512  0.20323026 -5.5499447 -1.4166964  2.95159827  0.94325321
#> [17,] -0.99926614 -0.88673316 -1.7227036 -1.7376607  4.17882331 -0.45509189
#> [18,]  1.26620104 -1.35517056 -4.1898037 -4.0741176  0.13342843 -2.37959946
#> [19,] -0.09741393 -0.56016236  2.4888181  1.0717510  0.60295079 -0.81681448
#> [20,] -0.76957388  0.04267854 -4.0172799 -3.6917960 -1.17062241 -0.16816184
#>              [,7]       [,8]        [,9]      [,10]      [,11]       [,12]
#>  [1,] -1.10951371  0.6222631  0.47593743 -3.0080226  2.4202532  0.31930243
#>  [2,] -0.19866350  0.7800527  0.37353174  0.5565255 -1.4881788  0.92589568
#>  [3,] -2.57726886 -5.5952136 -5.67597163 -0.8115165 -0.9728255 -4.17977295
#>  [4,] -3.35213669 -0.8600820  0.67946836 -4.3547601  1.1141970  0.95473935
#>  [5,]  0.48320219 -2.6948321  3.24225454 -3.5629398  4.1892918  0.07778905
#>  [6,] -0.08115338  0.3580289  1.08780011  0.4243304 -1.7966753  1.32476980
#>  [7,]  1.88847438 -0.5192815  0.66363509  1.8015421 -0.4084005 -0.21942540
#>  [8,] -0.51928150  4.0575882  0.49930031  0.6227474 -1.7792765  2.18749774
#>  [9,]  0.66363509  0.4993003  5.11079789 -2.4691545  1.4232707  2.95767958
#> [10,]  1.80154211  0.6227474 -2.46915452  6.4911189 -4.6111187 -1.03559447
#> [11,] -0.40840053 -1.7792765  1.42327071 -4.6111187  5.1245965 -0.67405923
#> [12,] -0.21942540  2.1874977  2.95767958 -1.0355945 -0.6740592  2.87059905
#> [13,] -0.86792254 -0.1013811  0.54400613 -2.8073587  3.5450428 -0.38495703
#> [14,] -2.74243137 -0.1094670  0.88339615 -6.7012490  3.3321615  0.95186069
#> [15,]  0.81711497  0.2216832  0.14090389 -0.3433361  1.4867044 -0.41436016
#> [16,]  1.58740460  0.3527846  4.94154808 -2.4570223  1.7165243  2.62571884
#> [17,]  2.03004478 -3.2378631  2.13465753  0.2749298  1.8877442 -0.84140503
#> [18,]  0.92533729  0.5176357 -1.18754700 -0.2005181  2.1517370 -1.35291155
#> [19,] -0.51953764 -1.5512974 -1.56914539  0.4611332  0.5125772 -1.63947636
#> [20,]  1.45034539  1.1825271  0.02425492  2.2475503 -1.2442342  0.23852424
#>            [,13]      [,14]      [,15]      [,16]       [,17]       [,18]
#>  [1,]  2.2330840  2.9701040  0.6961530  0.3954951 -0.99926614  1.26620104
#>  [2,] -1.2453065  0.1092993 -0.7754678  0.2032303 -0.88673316 -1.35517056
#>  [3,] -3.4684265  5.7044306 -3.4265740 -5.5499447 -1.72270357 -4.18980374
#>  [4,]  0.8612286  5.7993972 -2.5327763 -1.4166964 -1.73766073 -4.07411757
#>  [5,]  2.5144419  1.2442506  0.7052079  2.9515983  4.17882331  0.13342843
#>  [6,] -2.0011618  0.4233556 -1.2453874  0.9432532 -0.45509189 -2.37959946
#>  [7,] -0.8679225 -2.7424314  0.8171150  1.5874046  2.03004478  0.92533729
#>  [8,] -0.1013811 -0.1094670  0.2216832  0.3527846 -3.23786312  0.51763568
#>  [9,]  0.5440061  0.8833961  0.1409039  4.9415481  2.13465753 -1.18754700
#> [10,] -2.8073587 -6.7012490 -0.3433361 -2.4570223  0.27492983 -0.20051813
#> [11,]  3.5450428  3.3321615  1.4867044  1.7165243  1.88774420  2.15173700
#> [12,] -0.3849570  0.9518607 -0.4143602  2.6257188 -0.84140503 -1.35291155
#> [13,]  4.0626188  0.9775981  1.1669455 -0.4234292  0.78581591  1.93829192
#> [14,]  0.9775981  9.4890908 -0.4285345  1.6379416 -3.10494440 -0.48883914
#> [15,]  1.1669455 -0.4285345  1.5666726  1.0458193  0.62177136  2.54561099
#> [16,] -0.4234292  1.6379416  1.0458193  6.5006701  2.03428722  0.31323285
#> [17,]  0.7858159 -3.1049444  0.6217714  2.0342872  5.52288578  0.01256316
#> [18,]  1.9382919 -0.4888391  2.5456110  0.3132329  0.01256316  4.58060857
#> [19,]  1.2104091 -1.2300612 -0.3863602 -2.7538974  1.00954490 -0.28704537
#> [20,] -0.8615589 -2.6612145  0.9243385  0.9164802  0.16362712  1.42560072
#>             [,19]       [,20]
#>  [1,] -0.09741393 -0.76957388
#>  [2,] -0.56016236  0.04267854
#>  [3,]  2.48881809 -4.01727992
#>  [4,]  1.07175101 -3.69179598
#>  [5,]  0.60295079 -1.17062241
#>  [6,] -0.81681448 -0.16816184
#>  [7,] -0.51953764  1.45034539
#>  [8,] -1.55129742  1.18252715
#>  [9,] -1.56914539  0.02425492
#> [10,]  0.46113320  2.24755026
#> [11,]  0.51257717 -1.24423424
#> [12,] -1.63947636  0.23852424
#> [13,]  1.21040912 -0.86155893
#> [14,] -1.23006120 -2.66121451
#> [15,] -0.38636018  0.92433853
#> [16,] -2.75389739  0.91648023
#> [17,]  1.00954490  0.16362712
#> [18,] -0.28704537  1.42560072
#> [19,]  2.08766361 -0.93396219
#> [20,] -0.93396219  1.88049631
#> 
#> , , 2
#> 
#>              [,1]        [,2]       [,3]       [,4]        [,5]        [,6]
#>  [1,]  3.44413887  1.26959264  1.2378825  1.0950183  0.76673151  1.87358683
#>  [2,]  1.26959264  1.77858542 -0.2198115  0.7659096  0.11335942  1.55462979
#>  [3,]  1.23788250 -0.21981149  4.1035387  3.2429428  2.69337670  1.68447675
#>  [4,]  1.09501834  0.76590960  3.2429428  8.9728584  2.89299031 -0.40475285
#>  [5,]  0.76673151  0.11335942  2.6933767  2.8929903  2.19286845  0.64901278
#>  [6,]  1.87358683  1.55462979  1.6844767 -0.4047529  0.64901278  3.99681767
#>  [7,]  0.79849451  0.23015757 -1.1943872 -3.9316539 -0.97237624  0.72029005
#>  [8,] -1.04228519  1.08281735 -0.8804542  1.2031090 -0.55106578  0.66432124
#>  [9,] -1.18617945 -0.71967980 -0.3808410 -1.1080962 -0.53625095 -0.36409009
#> [10,] -0.41611808 -0.05076391  4.8574687  3.6759255  2.69991251  3.40773920
#> [11,] -0.59057988  0.61934323 -2.2550560 -2.3371620 -1.49219136 -0.03014889
#> [12,] -0.19811542 -0.50990550  2.4358073 -0.5748421  2.21529820  0.53747710
#> [13,] -1.66937433 -0.04172636  0.9204773 -1.4043704  2.00517488 -0.54909423
#> [14,]  1.15529340  2.52175529 -0.3816033  0.1106376 -0.86786024  3.96893829
#> [15,]  1.40295050  0.11544473  1.9403566  0.6123041  0.58729133  2.21242869
#> [16,]  1.86031402  1.24038189  0.6790198  3.4904505 -0.23026195  1.84430652
#> [17,] -0.07978953 -2.17926040  1.8398444 -1.9292950  0.65064488 -0.29834889
#> [18,] -1.06096135  1.05970034 -4.2006820 -0.1773063 -2.32179782 -2.14757587
#> [19,] -0.62178843 -0.09315525 -1.4881513 -1.7080956 -0.01356917 -2.02519402
#> [20,] -2.55773181 -0.90948119  1.4175818  2.7653880  1.05920453 -0.73711153
#>              [,7]       [,8]        [,9]       [,10]       [,11]      [,12]
#>  [1,]  0.79849451 -1.0422852 -1.18617945 -0.41611808 -0.59057988 -0.1981154
#>  [2,]  0.23015757  1.0828173 -0.71967980 -0.05076391  0.61934323 -0.5099055
#>  [3,] -1.19438722 -0.8804542 -0.38084096  4.85746871 -2.25505601  2.4358073
#>  [4,] -3.93165394  1.2031090 -1.10809622  3.67592549 -2.33716199 -0.5748421
#>  [5,] -0.97237624 -0.5510658 -0.53625095  2.69991251 -1.49219136  2.2152982
#>  [6,]  0.72029005  0.6643212 -0.36409009  3.40773920 -0.03014889  0.5374771
#>  [7,]  2.33037081 -1.0932861 -0.05091026 -2.39257933  0.97854613  0.6233516
#>  [8,] -1.09328610  2.6252769  0.18420341  1.35806231  0.89147509 -1.7263066
#>  [9,] -0.05091026  0.1842034  0.63766708  0.45717851  0.17839517 -0.1680172
#> [10,] -2.39257933  1.3580623  0.45717851  9.40839733 -2.12302461  2.0020784
#> [11,]  0.97854613  0.8914751  0.17839517 -2.12302461  1.56507368 -1.0485099
#> [12,]  0.62335164 -1.7263066 -0.16801722  2.00207835 -1.04850987  5.1068123
#> [13,]  0.89755359 -0.9881515 -0.20275907  0.26436812  0.04971903  6.2644953
#> [14,]  0.01263142  3.0692920 -0.18071099  2.71536171  1.26938398 -2.6977812
#> [15,] -0.13253457 -0.1922637 -0.01416424  3.03971927 -0.90560522 -0.2684682
#> [16,] -1.52585035  1.7864532 -0.39613054  2.15540818 -0.39610824 -3.9654463
#> [17,]  0.74045099 -2.7048943  0.63311926  1.19079398 -1.41847302  2.2535914
#> [18,] -0.15109842  2.0981594 -0.11417940 -4.86958674  2.19656123 -3.6174622
#> [19,]  1.21456995 -1.1465792 -0.35286530 -3.93909742  0.74335629  2.1846744
#> [20,] -2.44874647  1.4193349  0.76191878  4.05557701 -0.89153951  0.7043839
#>             [,13]       [,14]       [,15]       [,16]       [,17]      [,18]
#>  [1,] -1.66937433  1.15529340  1.40295050  1.86031402 -0.07978953 -1.0609614
#>  [2,] -0.04172636  2.52175529  0.11544473  1.24038189 -2.17926040  1.0597003
#>  [3,]  0.92047729 -0.38160326  1.94035657  0.67901979  1.83984440 -4.2006820
#>  [4,] -1.40437035  0.11063761  0.61230413  3.49045049 -1.92929496 -0.1773063
#>  [5,]  2.00517488 -0.86786024  0.58729133 -0.23026195  0.65064488 -2.3217978
#>  [6,] -0.54909423  3.96893829  2.21242869  1.84430652 -0.29834889 -2.1475759
#>  [7,]  0.89755359  0.01263142 -0.13253457 -1.52585035  0.74045099 -0.1510984
#>  [8,] -0.98815150  3.06929203 -0.19226366  1.78645317 -2.70489430  2.0981594
#>  [9,] -0.20275907 -0.18071099 -0.01416424 -0.39613054  0.63311926 -0.1141794
#> [10,]  0.26436812  2.71536171  3.03971927  2.15540818  1.19079398 -4.8695867
#> [11,]  0.04971903  1.26938398 -0.90560522 -0.39610824 -1.41847302  2.1965612
#> [12,]  6.26449529 -2.69778122 -0.26846824 -3.96544627  2.25359137 -3.6174622
#> [13,] 10.06865714 -3.42688940 -2.41198968 -6.46412016  0.69815234 -1.6837962
#> [14,] -3.42688940  6.91829103  1.83446232  4.23906468 -3.17493052  1.1660420
#> [15,] -2.41198968  1.83446232  2.30274016  2.20576953  1.11181292 -2.4233185
#> [16,] -6.46412016  4.23906468  2.20576953  5.75615959 -1.92088694  0.8308740
#> [17,]  0.69815234 -3.17493052  1.11181292 -1.92088694  4.41260418 -3.9581689
#> [18,] -1.68379623  1.16604201 -2.42331854  0.83087405 -3.95816891  5.9389430
#> [19,]  4.74798088 -3.19917649 -2.55553160 -4.02197321 -0.15256591  1.2928683
#> [20,]  1.12311180 -0.36933165 -0.09032065 -0.00710502  0.02470990 -0.7336649
#>             [,19]       [,20]
#>  [1,] -0.62178843 -2.55773181
#>  [2,] -0.09315525 -0.90948119
#>  [3,] -1.48815132  1.41758184
#>  [4,] -1.70809559  2.76538796
#>  [5,] -0.01356917  1.05920453
#>  [6,] -2.02519402 -0.73711153
#>  [7,]  1.21456995 -2.44874647
#>  [8,] -1.14657923  1.41933491
#>  [9,] -0.35286530  0.76191878
#> [10,] -3.93909742  4.05557701
#> [11,]  0.74335629 -0.89153951
#> [12,]  2.18467440  0.70438386
#> [13,]  4.74798088  1.12311180
#> [14,] -3.19917649 -0.36933165
#> [15,] -2.55553160 -0.09032065
#> [16,] -4.02197321 -0.00710502
#> [17,] -0.15256591  0.02470990
#> [18,]  1.29286827 -0.73366491
#> [19,]  4.05282600 -1.04002514
#> [20,] -1.04002514  4.00668140
#>