Nonsingular Wishart Helper Function

An expansion of R's 'stats' random wishart matrix generation. This package allows the user to generate singular, Uhlig and Harald (1994) <doi:10.1214/aos/1176325375>, and pseudo wishart, Diaz-Garcia, et al.(1997) <doi:10.1006/jmva.1997.1689>, matrices. In addition the user can generate wishart matrices with fractional degrees of freedom, Adhikari (2008) <doi:10.1061/(ASCE)0733-9399(2008)134:12(1029)>, commonly used in volatility modeling. Users can also use this package to create random covariance matrices.

Generate n random matrices, distributed according to the Wishart distribution with parameters Sigma and df, W_p(Sigma, df).

NonsingularWishart(df, Sigma, covariance = FALSE)

Arguments

df
numeric parameter, “degrees of freedom”.

Sigma
positive definite (\(p\times p\)) “scale” matrix, the matrix parameter of the distribution.

covariance
logical on whether a covariance matrix should be generated

Value

A numeric array of dimension p * p * n, where each array is a positive semidefinite matrix, a realization of the Wishart distribution W_p(Sigma, df)

Details

If X_1, ..., X_m is a sample of m independent multivariate Gaussians with mean vector 0, and covariance matrix Sigma, the distribution of M = X'X is W_p(Sigma, m).

Examples

NonsingularWishart(20, diag(1,5))

#>           [,1]       [,2]      [,3]      [,4]       [,5]
#> [1,] 21.647018  2.7918198 -6.000800 -2.523869 -7.1136878
#> [2,]  2.791820 10.7008502 -1.180177  0.390330 -0.1539626
#> [3,] -6.000800 -1.1801772 19.023804 -2.014090 -3.5254185
#> [4,] -2.523869  0.3903300 -2.014090 19.005544  4.9197783
#> [5,] -7.113688 -0.1539626 -3.525418  4.919778 17.4912123
#> attr(,"df")
#> ~20
#> <environment: namespace:rWishart>