Random Nonsingular Wishart Matrix

Generate n random matrices, distributed according to the Wishart distribution with parameters Sigma and df, W_p(Sigma, df).

rNonsingularWishart(n, df, Sigma, covariance = FALSE, simplify = "array")

Arguments

n
integer: the number of replications.

df
numeric parameter, “degrees of freedom”.

Sigma
positive definite (\(p\times p\)) “scale” matrix, the matrix parameter of the distribution.

covariance
logical on whether a covariance matrix should be generated

simplify
logical or character string; should the result be simplified to a vector, matrix or higher dimensional array if possible? For sapply it must be named and not abbreviated. The default value, TRUE, returns a vector or matrix if appropriate, whereas if simplify = "array" the result may be an array of “rank” (\(=\)length(dim(.))) one higher than the result of FUN(X[[i]]).

Value

A numeric array of dimension p * p * n, where each array is a positive semidefinite matrix, a realization of the Wishart distribution W_p(Sigma, df)

Details

If X_1, ..., X_m is a sample of m independent multivariate Gaussians with mean vector 0, and covariance matrix Sigma, the distribution of M = X'X is W_p(Sigma, m).

Examples

rNonsingularWishart(2, 20, diag(1, 5))

#> , , 1
#> 
#>           [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] 21.295048  1.7198774  2.7321114  3.1580862  2.6844694
#> [2,]  1.719877 27.5115368  6.1314136 -0.5960628 -8.5016441
#> [3,]  2.732111  6.1314136 10.7232983 -5.1603065  0.9389226
#> [4,]  3.158086 -0.5960628 -5.1603065 17.6416121 -3.8220035
#> [5,]  2.684469 -8.5016441  0.9389226 -3.8220035 14.3026934
#> 
#> , , 2
#> 
#>           [,1]       [,2]      [,3]       [,4]      [,5]
#> [1,] 33.815936  5.4101845 -7.440097 -8.3990749  6.442904
#> [2,]  5.410185 13.2196472 -4.242478 -0.4458656 -0.560210
#> [3,] -7.440097 -4.2424780 16.400520  1.3271250 -6.269885
#> [4,] -8.399075 -0.4458656  1.327125 25.6167644 -4.074066
#> [5,]  6.442904 -0.5602100 -6.269885 -4.0740663 23.254052
#>