Random Psuedo Wishart Matrix

Random Psuedo Wishart Matrix

rPsuedoWishart(n, df, Sigma, covariance = FALSE, simplify = "array")

Arguments

n

integer: the number of replications.

df

degrees of freedom

Sigma

Covariance matrix

covariance

should covariance matrix be generated (X / df)

simplify

logical or character string; should the result be simplified to a vector, matrix or higher dimensional array if possible? For sapply it must be named and not abbreviated. The default value, TRUE, returns a vector or matrix if appropriate, whereas if simplify = "array" the result may be an array of “rank” (\(=\)length(dim(.))) one higher than the result of FUN(X[[i]]).

Value

list or array of random wishart matrices

Examples

rPsuedoWishart(2, 5, diag(1, 20))

#> , , 1
#> 
#>             [,1]       [,2]        [,3]        [,4]       [,5]       [,6]
#>  [1,]  5.1160061 -2.6929421 -2.71234827 -0.17761570 -2.2289724 -1.6878254
#>  [2,] -2.6929421  6.1745550 -1.01180499 -1.56108505  0.8118241 -0.1347258
#>  [3,] -2.7123483 -1.0118050  5.20132781  0.43655216  0.2357286  1.9712222
#>  [4,] -0.1776157 -1.5610851  0.43655216  1.05046532  0.5908006  0.6360233
#>  [5,] -2.2289724  0.8118241  0.23572864  0.59080061  4.7872066 -2.5137160
#>  [6,] -1.6878254 -0.1347258  1.97122220  0.63602326 -2.5137160  4.4368498
#>  [7,]  0.1593877  4.1505765 -1.44145839 -0.61804375 -0.8152401  0.6501181
#>  [8,]  2.6167445 -4.7213553  0.11306737  1.47182412  0.7013411 -1.2092153
#>  [9,] -4.1144135 -1.5397839  6.66825528  0.80276315 -0.1265052  3.6176977
#> [10,] -4.0470477  4.5249816  2.67938674 -1.83232390  1.6484255 -0.7124830
#> [11,] -2.3580765  2.0101747  2.20105931 -0.92369127 -2.0430498  2.6965575
#> [12,] -1.0655294  1.8545598  0.09995559 -0.73364298 -2.4375440  2.4652629
#> [13,]  0.6584534 -2.6220967 -1.34944839  1.81945614  0.2378020  1.2488817
#> [14,]  2.4435882 -1.5192450 -2.68803204  0.01688647 -2.1588966  0.3307828
#> [15,]  0.9283063 -2.2470876 -1.65245737  1.45211517  2.2294814 -1.3445433
#> [16,] -2.8636238  5.2195700 -1.30231135 -1.02930936 -1.2025817  2.5071409
#> [17,]  0.2392981 -0.6618040 -1.35679362  1.31899478  2.6959122 -1.2586293
#> [18,] -0.1593296  0.5382527 -0.42860330  1.06419379 -0.6551376  2.2727546
#> [19,]  1.0269138 -3.5124574  0.40895279  1.22422293  2.3738351 -2.0238298
#> [20,] -1.5664368 -1.4093012  2.20203346  1.35338401  3.1039635 -0.5537556
#>             [,7]       [,8]       [,9]      [,10]      [,11]       [,12]
#>  [1,]  0.1593877  2.6167445 -4.1144135 -4.0470477 -2.3580765 -1.06552936
#>  [2,]  4.1505765 -4.7213553 -1.5397839  4.5249816  2.0101747  1.85455984
#>  [3,] -1.4414584  0.1130674  6.6682553  2.6793867  2.2010593  0.09995559
#>  [4,] -0.6180437  1.4718241  0.8027632 -1.8323239 -0.9236913 -0.73364298
#>  [5,] -0.8152401  0.7013411 -0.1265052  1.6484255 -2.0430498 -2.43754403
#>  [6,]  0.6501181 -1.2092153  3.6176977 -0.7124830  2.6965575  2.46526294
#>  [7,]  6.9779241 -2.1789252 -2.7339212  0.1962164  0.0882854  0.80544399
#>  [8,] -2.1789252  4.6093109 -0.2366957 -4.1659831 -3.2318642 -2.82252749
#>  [9,] -2.7339212 -0.2366957  9.1948869  3.0635857  3.4984266  0.93458310
#> [10,]  0.1962164 -4.1659831  3.0635857  7.7726478  3.3700803  1.26291022
#> [11,]  0.0882854 -3.2318642  3.4984266  3.3700803  4.0380501  2.94275495
#> [12,]  0.8054440 -2.8225275  0.9345831  1.2629102  2.9427550  2.89723027
#> [13,] -0.8526182  2.3470182 -0.8942606 -4.8346856 -2.0018822 -0.67703644
#> [14,] -1.1525535  0.5723115 -2.7427408 -3.1284831 -0.4664393  0.87345148
#> [15,] -1.1432109  2.9023395 -2.0571995 -3.5299776 -3.3405401 -2.29274669
#> [16,]  3.2590957 -5.0875205 -0.7338392  2.5753731  3.2221498  3.62703933
#> [17,]  1.1127895  2.1558567 -2.0995172 -2.6637217 -3.2477477 -2.34197062
#> [18,]  3.7559897  0.2002857 -0.4480101 -2.8194608 -0.5734826  0.31403978
#> [19,] -2.8470357  3.7240292  0.1633172 -2.0886944 -2.9725844 -3.00157139
#> [20,] -0.7094295  2.1776243  2.4167463 -0.1522296 -1.6863689 -2.39862103
#>            [,13]       [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,]  0.6584534  2.44358816  0.9283063 -2.8636238  0.2392981 -0.1593296
#>  [2,] -2.6220967 -1.51924504 -2.2470876  5.2195700 -0.6618040  0.5382527
#>  [3,] -1.3494484 -2.68803204 -1.6524574 -1.3023113 -1.3567936 -0.4286033
#>  [4,]  1.8194561  0.01688647  1.4521152 -1.0293094  1.3189948  1.0641938
#>  [5,]  0.2378020 -2.15889656  2.2294814 -1.2025817  2.6959122 -0.6551376
#>  [6,]  1.2488817  0.33078283 -1.3445433  2.5071409 -1.2586293  2.2727546
#>  [7,] -0.8526182 -1.15255348 -1.1432109  3.2590957  1.1127895  3.7559897
#>  [8,]  2.3470182  0.57231146  2.9023395 -5.0875205  2.1558567  0.2002857
#>  [9,] -0.8942606 -2.74274084 -2.0571995 -0.7338392 -2.0995172 -0.4480101
#> [10,] -4.8346856 -3.12848313 -3.5299776  2.5753731 -2.6637217 -2.8194608
#> [11,] -2.0018822 -0.46643931 -3.3405401  3.2221498 -3.2477477 -0.5734826
#> [12,] -0.6770364  0.87345148 -2.2927467  3.6270393 -2.3419706  0.3140398
#> [13,]  4.3191202  1.71870293  3.1238763 -0.6762123  2.5109397  2.3136632
#> [14,]  1.7187029  3.17640895  0.7028075  0.3567008 -0.4952053 -0.1405235
#> [15,]  3.1238763  0.70280752  3.6680136 -2.2666783  3.2551747  0.9142630
#> [16,] -0.6762123  0.35670080 -2.2666783  6.8957193 -1.4293440  1.5343035
#> [17,]  2.5109397 -0.49520530  3.2551747 -1.4293440  3.8045795  2.0519627
#> [18,]  2.3136632 -0.14052346  0.9142630  1.5343035  2.0519627  4.2934788
#> [19,]  1.5960522 -0.29755708  2.8712656 -4.5271098  2.2036570 -0.8433151
#> [20,]  1.0370642 -2.30951615  1.9642864 -2.6708126  2.4659610  0.7380746
#>            [,19]      [,20]
#>  [1,]  1.0269138 -1.5664368
#>  [2,] -3.5124574 -1.4093012
#>  [3,]  0.4089528  2.2020335
#>  [4,]  1.2242229  1.3533840
#>  [5,]  2.3738351  3.1039635
#>  [6,] -2.0238298 -0.5537556
#>  [7,] -2.8470357 -0.7094295
#>  [8,]  3.7240292  2.1776243
#>  [9,]  0.1633172  2.4167463
#> [10,] -2.0886944 -0.1522296
#> [11,] -2.9725844 -1.6863689
#> [12,] -3.0015714 -2.3986210
#> [13,]  1.5960522  1.0370642
#> [14,] -0.2975571 -2.3095161
#> [15,]  2.8712656  1.9642864
#> [16,] -4.5271098 -2.6708126
#> [17,]  2.2036570  2.4659610
#> [18,] -0.8433151  0.7380746
#> [19,]  3.8600616  2.7047605
#> [20,]  2.7047605  3.8093588
#> 
#> , , 2
#> 
#>              [,1]        [,2]       [,3]       [,4]        [,5]       [,6]
#>  [1,]  6.50048408  3.35898912 -0.3207295  5.1877660  0.98233714  0.3638402
#>  [2,]  3.35898912  3.31530905 -1.2363603  3.2638494  0.14803112  0.3499761
#>  [3,] -0.32072947 -1.23636029  2.5441284  1.6087439 -0.74644629  0.7345091
#>  [4,]  5.18776600  3.26384937  1.6087439  8.6426670 -0.71915096  1.2508437
#>  [5,]  0.98233714  0.14803112 -0.7464463 -0.7191510  1.67756850 -0.1889220
#>  [6,]  0.36384020  0.34997611  0.7345091  1.2508437 -0.18892198  0.5236966
#>  [7,] -1.94117580  0.21670320 -1.5530903 -1.3689851 -0.42908005 -0.5156207
#>  [8,] -2.81465219 -2.25574920  3.1370089  2.1291862 -0.34237741  1.0655985
#>  [9,]  2.57664592  0.09860344  1.1961050  3.0674252 -2.11218656 -0.4257158
#> [10,] -0.84242765 -1.32662898  0.1760402 -1.2974579 -0.58679668 -0.6049084
#> [11,] -0.04324542  1.25754634  0.5465893  2.3735961 -0.01622408  1.0295533
#> [12,] -1.84513985 -2.84614053 -1.0550716 -7.4575664  2.47211605 -1.0019714
#> [13,]  2.91534099  0.03627867 -0.5491989 -0.5589280  1.90432402 -0.5097833
#> [14,] -0.90960750 -0.20713317  0.1889374 -0.8642396 -1.78906049 -0.1465509
#> [15,]  1.02217789 -3.11040249  1.7744799 -1.9262158  1.21629959 -0.5911365
#> [16,] -0.85365923 -0.48743250  0.7203520  0.5884735 -1.79176301 -0.1029773
#> [17,] -0.35158541  1.96080261 -1.6246748  0.3145391 -0.23960060  0.1529661
#> [18,] -4.41075005 -3.32596640  1.2549170 -4.3401586  0.19483565  0.0689369
#> [19,]  4.28715898  0.74479134  2.3854146  5.2643125 -0.56446406  0.6512410
#> [20,]  0.41605716  0.97599748 -1.9489964 -0.7594915  0.75822026 -0.5806638
#>             [,7]       [,8]        [,9]      [,10]       [,11]      [,12]
#>  [1,] -1.9411758 -2.8146522  2.57664592 -0.8424277 -0.04324542 -1.8451398
#>  [2,]  0.2167032 -2.2557492  0.09860344 -1.3266290  1.25754634 -2.8461405
#>  [3,] -1.5530903  3.1370089  1.19610499  0.1760402  0.54658933 -1.0550716
#>  [4,] -1.3689851  2.1291862  3.06742521 -1.2974579  2.37359613 -7.4575664
#>  [5,] -0.4290800 -0.3423774 -2.11218656 -0.5867967 -0.01622408  2.4721160
#>  [6,] -0.5156207  1.0655985 -0.42571585 -0.6049084  1.02955331 -1.0019714
#>  [7,]  2.3324732 -0.3457974 -1.00782659  0.1065259  0.23325685 -1.1286386
#>  [8,] -0.3457974  8.7943872 -2.50041699 -0.8016265  3.02950787 -2.8709390
#>  [9,] -1.0078266 -2.5004170  8.88988194  2.6641167 -3.23889770 -3.1987664
#> [10,]  0.1065259 -0.8016265  2.66411667  1.6058548 -2.02366825  0.7797864
#> [11,]  0.2332568  3.0295079 -3.23889770 -2.0236682  3.32621082 -2.8301852
#> [12,] -1.1286386 -2.8709390 -3.19876636  0.7797864 -2.83018518 10.1375737
#> [13,] -1.7375367 -2.5881652  0.74773477  0.3997161 -2.01678958  3.9863887
#> [14,]  0.2980829 -2.2538811  3.05753874  1.1648880 -1.28673762 -0.7340675
#> [15,] -3.0973609 -0.3894705  3.18654826  2.1352687 -3.72515404  5.7044102
#> [16,]  0.4307193 -0.1076739  3.45816489  1.2537956 -0.89520594 -2.3454912
#> [17,]  1.8260928 -0.8367658 -2.37679176 -1.2992984  1.82647304 -2.1073171
#> [18,] -0.2209200  2.9841707 -3.11522688  0.2676520  0.10693398  3.9242436
#> [19,] -2.8131671  0.4961624  5.22055574  0.5518115 -0.66333813 -2.7356393
#> [20,]  1.2792241 -1.7802774 -1.05096304 -0.2057007 -0.24300809  0.4080703
#>             [,13]       [,14]      [,15]      [,16]       [,17]      [,18]
#>  [1,]  2.91534099 -0.90960750  1.0221779 -0.8536592 -0.35158541 -4.4107501
#>  [2,]  0.03627867 -0.20713317 -3.1104025 -0.4874325  1.96080261 -3.3259664
#>  [3,] -0.54919889  0.18893744  1.7744799  0.7203520 -1.62467483  1.2549170
#>  [4,] -0.55892800 -0.86423960 -1.9262158  0.5884735  0.31453905 -4.3401586
#>  [5,]  1.90432402 -1.78906049  1.2162996 -1.7917630 -0.23960060  0.1948356
#>  [6,] -0.50978334 -0.14655094 -0.5911365 -0.1029773  0.15296614  0.0689369
#>  [7,] -1.73753667  0.29808292 -3.0973609  0.4307193  1.82609279 -0.2209200
#>  [8,] -2.58816519 -2.25388108 -0.3894705 -0.1076739 -0.83676577  2.9841707
#>  [9,]  0.74773477  3.05753874  3.1865483  3.4581649 -2.37679176 -3.1152269
#> [10,]  0.39971607  1.16488795  2.1352687  1.2537956 -1.29929836  0.2676520
#> [11,] -2.01678958 -1.28673762 -3.7251540 -0.8952059  1.82647304  0.1069340
#> [12,]  3.98638873 -0.73406747  5.7044102 -2.3454912 -2.10731712  3.9242436
#> [13,]  4.11183579 -1.16464555  4.5213242 -1.4727648 -1.94956096 -0.6831862
#> [14,] -1.16464555  2.94990363 -0.3117688  2.0799279  0.01718319 -0.2347523
#> [15,]  4.52132417 -0.31176876  9.0478946 -0.2073338 -4.98003708  1.8149693
#> [16,] -1.47276483  2.07992794 -0.2073338  2.2435528 -0.35494096 -0.4849327
#> [17,] -1.94956096  0.01718319 -4.9800371 -0.3549410  2.99242127 -1.0877412
#> [18,] -0.68318615 -0.23475231  1.8149693 -0.4849327 -1.08774118  4.7396531
#> [19,]  1.66629724  0.34578079  3.4641622  1.0837767 -2.53145575 -2.3502863
#> [20,]  0.47927864 -0.65510739 -1.3831864 -0.7032899  1.21079502 -1.1407769
#>            [,19]      [,20]
#>  [1,]  4.2871590  0.4160572
#>  [2,]  0.7447913  0.9759975
#>  [3,]  2.3854146 -1.9489964
#>  [4,]  5.2643125 -0.7594915
#>  [5,] -0.5644641  0.7582203
#>  [6,]  0.6512410 -0.5806638
#>  [7,] -2.8131671  1.2792241
#>  [8,]  0.4961624 -1.7802774
#>  [9,]  5.2205557 -1.0509630
#> [10,]  0.5518115 -0.2057007
#> [11,] -0.6633381 -0.2430081
#> [12,] -2.7356393  0.4080703
#> [13,]  1.6662972  0.4792786
#> [14,]  0.3457808 -0.6551074
#> [15,]  3.4641622 -1.3831864
#> [16,]  1.0837767 -0.7032899
#> [17,] -2.5314558  1.2107950
#> [18,] -2.3502863 -1.1407769
#> [19,]  6.1863727 -1.7138992
#> [20,] -1.7138992  1.6454242
#>