Random Wishart Distributed Matrices

Random Wishart Distributed Matrices

rWishart(n, df, Sigma, covariance = FALSE, simplify = "array")

Arguments

n
integer: the number of replications.

df
degrees of freedom

Sigma
Covariance matrix

covariance
should covariance matrix be generated (X / df)

simplify
logical or character string; should the result be simplified to a vector, matrix or higher dimensional array if possible? For sapply it must be named and not abbreviated. The default value, TRUE, returns a vector or matrix if appropriate, whereas if simplify = "array" the result may be an array of “rank” (\(=\)length(dim(.))) one higher than the result of FUN(X[[i]]).

Value

list or array of random wishart matrices

Examples

rWishart(2, 5, diag(1, 20))

#> , , 1
#> 
#>              [,1]       [,2]         [,3]       [,4]         [,5]       [,6]
#>  [1,]  8.00883031  1.5564829 -1.610897218 -2.9940382  2.192787380  0.2012462
#>  [2,]  1.55648292 12.2607081  3.357048579 -0.8166983  1.671411978 -3.0299667
#>  [3,] -1.61089722  3.3570486  2.436525741  0.2746025  1.027824901 -2.0494444
#>  [4,] -2.99403822 -0.8166983  0.274602517  2.1729816 -1.186220584  2.5147654
#>  [5,]  2.19278738  1.6714120  1.027824901 -1.1862206  2.965651307 -1.1662323
#>  [6,]  0.20124623 -3.0299667 -2.049444403  2.5147654 -1.166232279  7.2882507
#>  [7,]  1.65636264  0.4387346  0.168558541 -1.5686967  0.007219891 -2.8178298
#>  [8,] -0.08673388 -7.3508953 -0.420977292 -0.3828370  1.351980636 -0.3924400
#>  [9,]  2.95062377  0.2889932 -0.005481763 -0.1964436  1.577727572  1.8326932
#> [10,]  1.26816064 -3.1783091 -1.828637399  1.0153212  0.829056911  4.8865725
#> [11,]  3.62810716 -0.7425149 -0.645160567 -2.3332920  1.151668070 -2.0945190
#> [12,] -1.39349397  5.8106689  2.199679107 -1.3072110  0.544636995 -5.5550890
#> [13,]  1.13259869  1.6007067  1.542804182  0.8990458  2.864694453  2.3353133
#> [14,] -0.13760137 -3.6245893 -2.102175669 -0.7957764 -1.161498878 -0.6484262
#> [15,] -2.43199359 -0.5864736  1.062521133  1.1275600 -0.072308956  0.2814173
#> [16,] -0.88145567  3.1887223 -0.062922450  1.2537225 -2.383129992  1.6064085
#> [17,]  3.13402823 -1.3982193 -1.040747241 -0.9234875  0.750671280  1.0803797
#> [18,] -1.51950696 -0.1870159 -0.447455836  1.2923419 -1.904117990  1.7821584
#> [19,] -1.52549128  3.5092063  0.566581445  1.8695757 -1.628043582  2.4035449
#> [20,]  0.16656317 -9.3936482 -4.227118085  0.7149370 -3.849878387  3.7252361
#>               [,7]        [,8]         [,9]      [,10]      [,11]      [,12]
#>  [1,]  1.656362637 -0.08673388  2.950623767  1.2681606  3.6281072 -1.3934940
#>  [2,]  0.438734570 -7.35089533  0.288993150 -3.1783091 -0.7425149  5.8106689
#>  [3,]  0.168558541 -0.42097729 -0.005481763 -1.8286374 -0.6451606  2.1996791
#>  [4,] -1.568696672 -0.38283697 -0.196443648  1.0153212 -2.3332920 -1.3072110
#>  [5,]  0.007219891  1.35198064  1.577727572  0.8290569  1.1516681  0.5446370
#>  [6,] -2.817829754 -0.39243995  1.832693159  4.8865725 -2.0945190 -5.5550890
#>  [7,]  2.317940281  1.04626178  0.111040463 -2.2904692  1.7516220  1.2366284
#>  [8,]  1.046261776 10.24688232  2.776940623  3.2207761 -0.1727407 -2.4309085
#>  [9,]  0.111040463  2.77694062  3.659106078  3.0182045 -0.5081847 -1.8359261
#> [10,] -2.290469247  3.22077606  3.018204541  6.9966985 -2.6104712 -3.1659956
#> [11,]  1.751622009 -0.17274067 -0.508184708 -2.6104712  4.6959319 -0.7444298
#> [12,]  1.236628376 -2.43090855 -1.835926117 -3.1659956 -0.7444298  6.7086907
#> [13,] -1.168612935  4.38020565  5.004621418  5.1272937 -3.1204373 -1.0408896
#> [14,]  0.327483746  1.77615032 -1.122028143  1.5469819 -0.3362594  0.7733143
#> [15,] -0.767286414 -0.78217542 -1.280067791 -1.8950304  0.5825680 -1.0827840
#> [16,] -0.308465081 -4.87449983 -0.852746730 -1.3213753 -1.4955121  0.4000562
#> [17,]  0.292195619 -0.72410312  0.221899025 -0.7575956  3.4956078 -2.9807223
#> [18,] -0.612166947 -2.19927243 -0.955355610 -0.5243488 -0.9050231 -0.9448256
#> [19,] -1.217958985 -4.74489072 -0.450910840 -0.6247570 -1.9800786 -0.1268534
#> [20,]  0.785646960  4.74935291  0.475682062  2.9634053 -0.5822306 -4.8473238
#>            [,13]      [,14]       [,15]       [,16]       [,17]       [,18]
#>  [1,]  1.1325987 -0.1376014 -2.43199359 -0.88145567  3.13402823 -1.51950696
#>  [2,]  1.6007067 -3.6245893 -0.58647360  3.18872229 -1.39821930 -0.18701591
#>  [3,]  1.5428042 -2.1021757  1.06252113 -0.06292245 -1.04074724 -0.44745584
#>  [4,]  0.8990458 -0.7957764  1.12756000  1.25372249 -0.92348752  1.29234194
#>  [5,]  2.8646945 -1.1614989 -0.07230896 -2.38312999  0.75067128 -1.90411799
#>  [6,]  2.3353133 -0.6484262  0.28141730  1.60640846  1.08037975  1.78215844
#>  [7,] -1.1686129  0.3274837 -0.76728641 -0.30846508  0.29219562 -0.61216695
#>  [8,]  4.3802057  1.7761503 -0.78217542 -4.87449983 -0.72410312 -2.19927243
#>  [9,]  5.0046214 -1.1220281 -1.28006779 -0.85274673  0.22189902 -0.95535561
#> [10,]  5.1272937  1.5469819 -1.89503036 -1.32137529 -0.75759562 -0.52434877
#> [11,] -3.1204373 -0.3362594  0.58256797 -1.49551208  3.49560782 -0.90502310
#> [12,] -1.0408896  0.7733143 -1.08278397  0.40005620 -2.98072234 -0.94482563
#> [13,]  9.3608151 -2.0555408 -1.19962534 -1.71266838 -1.79508398 -1.64356378
#> [14,] -2.0555408  4.3810535 -2.08725962 -0.93896908 -1.24596445 -0.31209605
#> [15,] -1.1996253 -2.0872596  2.86047630 -0.01940350  1.14928668  0.67086623
#> [16,] -1.7126684 -0.9389691 -0.01940350  4.22494988 -0.61544306  2.24202782
#> [17,] -1.7950840 -1.2459645  1.14928668 -0.61544306  3.82636362 -0.02833014
#> [18,] -1.6435638 -0.3120961  0.67086623  2.24202782 -0.02833014  1.68239895
#> [19,] -0.2621857 -2.0235721  0.79911345  3.78775332 -0.51334125  2.10564825
#> [20,] -2.0244506  4.3029778 -1.89804477  0.47885719  0.10634409  1.52575135
#>            [,19]      [,20]
#>  [1,] -1.5254913  0.1665632
#>  [2,]  3.5092063 -9.3936482
#>  [3,]  0.5665814 -4.2271181
#>  [4,]  1.8695757  0.7149370
#>  [5,] -1.6280436 -3.8498784
#>  [6,]  2.4035449  3.7252361
#>  [7,] -1.2179590  0.7856470
#>  [8,] -4.7448907  4.7493529
#>  [9,] -0.4509108  0.4756821
#> [10,] -0.6247570  2.9634053
#> [11,] -1.9800786 -0.5822306
#> [12,] -0.1268534 -4.8473238
#> [13,] -0.2621857 -2.0244506
#> [14,] -2.0235721  4.3029778
#> [15,]  0.7991135 -1.8980448
#> [16,]  3.7877533  0.4788572
#> [17,] -0.5133413  0.1063441
#> [18,]  2.1056483  1.5257514
#> [19,]  4.0773535 -1.1642829
#> [20,] -1.1642829 12.3308888
#> 
#> , , 2
#> 
#>              [,1]        [,2]        [,3]        [,4]       [,5]        [,6]
#>  [1,]  9.27840136  3.89581843  0.54370463 -1.69751513 -2.5127576  0.09375841
#>  [2,]  3.89581843  3.92161113 -0.98196740 -0.38743399 -2.5720454  2.60521606
#>  [3,]  0.54370463 -0.98196740  4.31876332 -0.05763187  0.9223474 -4.26451109
#>  [4,] -1.69751513 -0.38743399 -0.05763187  1.86169666  0.3189374  0.62517450
#>  [5,] -2.51275763 -2.57204540  0.92234735  0.31893742  3.2808361 -3.02604121
#>  [6,]  0.09375841  2.60521606 -4.26451109  0.62517450 -3.0260412  6.29414713
#>  [7,] -2.83715187 -2.57659435  0.21662920  0.27090037 -0.8011167  0.58569481
#>  [8,]  5.89704840  4.13034339 -3.42592050  0.96936041 -0.9958142  3.29370575
#>  [9,] -2.00276636  1.65580798  0.54903827  1.62801035 -3.2089873  2.89824094
#> [10,] -0.69361839 -1.53129898  1.51007540  0.86811768  0.1570878 -0.82478844
#> [11,]  3.27293965 -0.08099955  0.64383504 -0.88709822 -1.1007335 -0.56500580
#> [12,]  1.55781746 -1.03546134 -0.19897196 -1.13311915 -0.5843280  0.10601021
#> [13,] -1.60606291 -3.44479589  1.20478348 -1.87751112  3.1536448 -4.32225966
#> [14,]  0.42490492  0.98311182  0.24496546 -0.42328216 -2.4304058  0.96341483
#> [15,]  1.61092215 -0.14184673  3.39447852 -3.07515632  0.8128487 -4.87825425
#> [16,] -1.94691974 -0.47926931 -0.73383085 -1.06080544 -2.0692726  2.04104785
#> [17,] -3.36835940  0.51744716  1.35949280  0.46631971 -0.9071069  0.60186711
#> [18,]  1.67323792  3.71858371 -2.97590390 -0.46159250 -3.2376103  4.62691826
#> [19,]  1.75963713 -0.83381443  1.13310012 -0.36362484  1.6899999 -2.46171460
#> [20,]  2.19040036 -1.63999577  0.81749055 -2.94016070  0.5665212 -2.95466321
#>             [,7]       [,8]       [,9]      [,10]       [,11]      [,12]
#>  [1,] -2.8371519  5.8970484 -2.0027664 -0.6936184  3.27293965  1.5578175
#>  [2,] -2.5765943  4.1303434  1.6558080 -1.5312990 -0.08099955 -1.0354613
#>  [3,]  0.2166292 -3.4259205  0.5490383  1.5100754  0.64383504 -0.1989720
#>  [4,]  0.2709004  0.9693604  1.6280104  0.8681177 -0.88709822 -1.1331192
#>  [5,] -0.8011167 -0.9958142 -3.2089873  0.1570878 -1.10073350 -0.5843280
#>  [6,]  0.5856948  3.2937058  2.8982409 -0.8247884 -0.56500580  0.1060102
#>  [7,]  5.7605689 -5.6700593  2.3284919  2.6954610  1.71160022  2.8730723
#>  [8,] -5.6700593 12.3483828 -2.4689810 -2.5002226 -0.13451613 -1.8916231
#>  [9,]  2.3284919 -2.4689810  8.0333332  0.8639556 -0.82745251 -1.6808799
#> [10,]  2.6954610 -2.5002226  0.8639556  2.7425427  1.11047608  1.6654216
#> [11,]  1.7116002 -0.1345161 -0.8274525  1.1104761  2.95981352  2.5933246
#> [12,]  2.8730723 -1.8916231 -1.6808799  1.6654216  2.59332456  3.7612389
#> [13,]  0.5961694 -3.9493104 -5.0093206 -0.2749185  0.58578748  1.4892266
#> [14,]  1.6422172 -1.6824276  3.8777097 -0.5971821  0.96134792 -0.3516846
#> [15,] -1.5581253 -4.3405642 -1.7678566 -1.8026761  0.53285142 -0.3491893
#> [16,]  4.1893490 -5.4328551  2.9729823  0.7236081  0.86404632  2.1280228
#> [17,]  0.9210737 -4.6591358  4.8076277  0.3432786 -2.11470346 -1.6627396
#> [18,] -1.4134159  3.3559603  2.9986288 -2.5451234 -0.68387142 -1.3556634
#> [19,] -1.1758152  1.1084011 -3.6086422  0.5310413  0.77909849  0.8552380
#> [20,]  1.3589846 -2.8618398 -3.9792888 -0.4878902  2.86324059  2.9648131
#>            [,13]      [,14]      [,15]      [,16]      [,17]      [,18]
#>  [1,] -1.6060629  0.4249049  1.6109222 -1.9469197 -3.3683594  1.6732379
#>  [2,] -3.4447959  0.9831118 -0.1418467 -0.4792693  0.5174472  3.7185837
#>  [3,]  1.2047835  0.2449655  3.3944785 -0.7338308  1.3594928 -2.9759039
#>  [4,] -1.8775111 -0.4232822 -3.0751563 -1.0608054  0.4663197 -0.4615925
#>  [5,]  3.1536448 -2.4304058  0.8128487 -2.0692726 -0.9071069 -3.2376103
#>  [6,] -4.3222597  0.9634148 -4.8782542  2.0410479  0.6018671  4.6269183
#>  [7,]  0.5961694  1.6422172 -1.5581253  4.1893490  0.9210737 -1.4134159
#>  [8,] -3.9493104 -1.6824276 -4.3405642 -5.4328551 -4.6591358  3.3559603
#>  [9,] -5.0093206  3.8777097 -1.7678566  2.9729823  4.8076277  2.9986288
#> [10,] -0.2749185 -0.5971821 -1.8026761  0.7236081  0.3432786 -2.5451234
#> [11,]  0.5857875  0.9613479  0.5328514  0.8640463 -2.1147035 -0.6838714
#> [12,]  1.4892266 -0.3516846 -0.3491893  2.1280228 -1.6627396 -1.3556634
#> [13,]  6.4117192 -1.2184867  4.5904421  0.1433986 -1.7682001 -3.5055364
#> [14,] -1.2184867  4.3359076  1.7839812  2.5716717  1.3517722  2.6054578
#> [15,]  4.5904421  1.7839812  9.3683616  0.6777830  1.1017890 -0.6562049
#> [16,]  0.1433986  2.5716717  0.6777830  5.3566047  2.4415949  1.4696375
#> [17,] -1.7682001  1.3517722  1.1017890  2.4415949  5.4565502  1.0756294
#> [18,] -3.5055364  2.6054578 -0.6562049  1.4696375  1.0756294  5.7102468
#> [19,]  1.9511652 -2.2666847  0.7512438 -2.2431891 -2.1960342 -2.5976299
#> [20,]  5.2274452  1.1094482  5.3031334  1.6433124 -2.7447928 -1.4198590
#>            [,19]      [,20]
#>  [1,]  1.7596371  2.1904004
#>  [2,] -0.8338144 -1.6399958
#>  [3,]  1.1331001  0.8174905
#>  [4,] -0.3636248 -2.9401607
#>  [5,]  1.6899999  0.5665212
#>  [6,] -2.4617146 -2.9546632
#>  [7,] -1.1758152  1.3589846
#>  [8,]  1.1084011 -2.8618398
#>  [9,] -3.6086422 -3.9792888
#> [10,]  0.5310413 -0.4878902
#> [11,]  0.7790985  2.8632406
#> [12,]  0.8552380  2.9648131
#> [13,]  1.9511652  5.2274452
#> [14,] -2.2666847  1.1094482
#> [15,]  0.7512438  5.3031334
#> [16,] -2.2431891  1.6433124
#> [17,] -2.1960342 -2.7447928
#> [18,] -2.5976299 -1.4198590
#> [19,]  2.3700009  1.3562118
#> [20,]  1.3562118  7.0985906
#>