Random Wishart Matrix Generation

An expansion of R's 'stats' random wishart matrix generation. This package allows the user to generate singular, Uhlig and Harald (1994) <doi:10.1214/aos/1176325375>, and pseudo wishart, Diaz-Garcia, et al.(1997) <doi:10.1006/jmva.1997.1689>, matrices. In addition the user can generate wishart matrices with fractional degrees of freedom, Adhikari (2008) <doi:10.1061/(ASCE)0733-9399(2008)134:12(1029)>, commonly used in volatility modeling. Users can also use this package to create random covariance matrices.

Generate n random matrices, distributed according to the Wishart distribution with parameters Sigma and df, W_p(Sigma, df).

rWishart(n, df, Sigma, covariance = FALSE, simplify = "array")

Arguments

n
integer: the number of replications.

df
numeric parameter, “degrees of freedom”.

Sigma
positive definite (\(p\times p\)) “scale” matrix, the matrix parameter of the distribution.

covariance
logical on whether a covariance matrix should be generated

simplify
logical or character string; should the result be simplified to a vector, matrix or higher dimensional array if possible? For sapply it must be named and not abbreviated. The default value, TRUE, returns a vector or matrix if appropriate, whereas if simplify = "array" the result may be an array of “rank” (\(=\)length(dim(.))) one higher than the result of FUN(X[[i]]).

Value

A numeric array of dimension p * p * n, where each array is a positive semidefinite matrix, a realization of the Wishart distribution W_p(Sigma, df)

Details

If X_1, ..., X_m is a sample of m independent multivariate Gaussians with mean vector 0, and covariance matrix Sigma, the distribution of M = X'X is W_p(Sigma, m).

Examples

rWishart(2, 5, diag(1, 20))

#> , , 1
#> 
#>              [,1]         [,2]        [,3]        [,4]       [,5]         [,6]
#>  [1,]  5.00425046 -0.270553781 -1.24762334  0.02946354  4.5189416  0.900080293
#>  [2,] -0.27055378  2.084698987 -0.24885346  0.30496693  0.2142017 -0.571724039
#>  [3,] -1.24762334 -0.248853465  2.70664376  1.10053928 -1.9657107 -0.873243548
#>  [4,]  0.02946354  0.304966933  1.10053928  2.21492131 -0.2876180 -0.868993157
#>  [5,]  4.51894161  0.214201670 -1.96571071 -0.28761801  4.5546636  0.696789293
#>  [6,]  0.90008029 -0.571724039 -0.87324355 -0.86899316  0.6967893  1.118316129
#>  [7,] -2.64341813 -2.101125554  1.70340059 -1.05564696 -2.9441997 -0.171171643
#>  [8,]  3.98788726 -0.003178642  0.83645614  1.68033441  3.2039849 -0.443343243
#>  [9,]  0.37685467 -2.378863305  0.66399604  0.49810260  0.0611505 -0.384670740
#> [10,]  1.71965888 -1.822539915 -0.08154119 -0.27802710  1.4852704  0.058258465
#> [11,]  1.56138242 -1.099216635  0.90653800  1.28145505  0.3602908  0.841991055
#> [12,] -2.83634683 -1.558322164 -0.87499173 -2.36380036 -2.1392729  0.432530589
#> [13,] -1.18442349  2.946137391  0.92728936  1.52452484 -0.7534573 -1.492827425
#> [14,]  2.37642831  0.665625507 -0.84651109  0.15714053  2.4087790  0.156132315
#> [15,] -1.32143683  0.642367247 -0.52825281 -0.74349106 -0.9046342  0.159897503
#> [16,] -2.00081571 -0.477525572 -0.11658182  0.24742690 -2.0401719  0.480662081
#> [17,] -3.70082836  0.323483005 -2.98457673 -0.43594353 -1.7386781 -0.503440937
#> [18,] -2.38379602 -0.691796234 -1.93852162 -1.49457524 -1.1544674  0.002576151
#> [19,] -5.05364861  2.346229340  1.04992124 -0.07920224 -4.8743534  0.112211849
#> [20,] -1.99447012 -2.295016177  2.20805357  0.27285144 -2.6423441 -0.424284550
#>              [,7]         [,8]       [,9]       [,10]       [,11]      [,12]
#>  [1,] -2.64341813  3.987887256  0.3768547  1.71965888  1.56138242 -2.8363468
#>  [2,] -2.10112555 -0.003178642 -2.3788633 -1.82253991 -1.09921664 -1.5583222
#>  [3,]  1.70340059  0.836456143  0.6639960 -0.08154119  0.90653800 -0.8749917
#>  [4,] -1.05564696  1.680334415  0.4981026 -0.27802710  1.28145505 -2.3638004
#>  [5,] -2.94419965  3.203984931  0.0611505  1.48527041  0.36029079 -2.1392729
#>  [6,] -0.17117164 -0.443343243 -0.3846707  0.05825846  0.84199106  0.4325306
#>  [7,]  4.96303632 -2.012985658  2.5700598  1.51467057 -0.57366109  3.8527515
#>  [8,] -2.01298566  5.226404581  1.1841707  1.70385382  1.80089970 -4.1538218
#>  [9,]  2.57005981  1.184170715  4.2141205  3.15093273  0.41798534  1.5478717
#> [10,]  1.51467057  1.703853816  3.1509327  3.01601295  0.10860004  0.9966792
#> [11,] -0.57366109  1.800899699  0.4179853  0.10860004  3.67695511 -2.3046076
#> [12,]  3.85275154 -4.153821830  1.5478717  0.99667925 -2.30460761  5.6637826
#> [13,] -2.51539514  0.500984188 -2.8623394 -2.71687329 -1.06428745 -2.9326926
#> [14,] -2.15491069  1.971002184 -0.5754599  0.26848149  0.17374492 -1.8961991
#> [15,]  0.01224757 -1.928576622 -1.3437015 -1.19780917 -0.96361141  1.0617990
#> [16,]  0.56342233 -2.371529728 -0.3585206 -1.20130973  0.93438687  1.1215934
#> [17,]  0.10987187 -4.941765059  0.4739439 -0.71085395 -3.44769941  4.4454140
#> [18,]  1.81705441 -3.685975323  1.2463115  0.65327258 -2.81273554  4.6539045
#> [19,] -0.29874747 -4.936679003 -5.1818855 -5.53891152 -0.01693155  0.1823178
#> [20,]  4.25541308 -0.595818432  3.2148704  1.74336783  0.70877046  2.2885913
#>            [,13]      [,14]       [,15]      [,16]      [,17]        [,18]
#>  [1,] -1.1844235  2.3764283 -1.32143683 -2.0008157 -3.7008284 -2.383796024
#>  [2,]  2.9461374  0.6656255  0.64236725 -0.4775256  0.3234830 -0.691796234
#>  [3,]  0.9272894 -0.8465111 -0.52825281 -0.1165818 -2.9845767 -1.938521615
#>  [4,]  1.5245248  0.1571405 -0.74349106  0.2474269 -0.4359435 -1.494575236
#>  [5,] -0.7534573  2.4087790 -0.90463421 -2.0401719 -1.7386781 -1.154467448
#>  [6,] -1.4928274  0.1561323  0.15989750  0.4806621 -0.5034409  0.002576151
#>  [7,] -2.5153951 -2.1549107  0.01224757  0.5634223  0.1098719  1.817054413
#>  [8,]  0.5009842  1.9710022 -1.92857662 -2.3715297 -4.9417651 -3.685975323
#>  [9,] -2.8623394 -0.5754599 -1.34370147 -0.3585206  0.4739439  1.246311514
#> [10,] -2.7168733  0.2684815 -1.19780917 -1.2013097 -0.7108539  0.653272582
#> [11,] -1.0642874  0.1737449 -0.96361141  0.9343869 -3.4476994 -2.812735543
#> [12,] -2.9326926 -1.8961991  1.06179905  1.1215934  4.4454140  4.653904470
#> [13,]  4.9986310  0.5262937  0.57355423 -0.4983960 -0.1179677 -1.733290196
#> [14,]  0.5262937  1.4616347 -0.42307366 -1.2163899 -1.2955730 -1.187494651
#> [15,]  0.5735542 -0.4230737  0.95006025  0.6596412  1.6481427  1.055698037
#> [16,] -0.4983960 -1.2163899  0.65964119  2.3974651  2.4083900  1.107469766
#> [17,] -0.1179677 -1.2955730  1.64814269  2.4083900 11.3292409  6.853606027
#> [18,] -1.7332902 -1.1874947  1.05569804  1.1074698  6.8536060  5.265165752
#> [19,]  3.7977208 -1.9006823  2.54162888  3.2595687  1.4392538 -0.409817805
#> [20,] -2.2337813 -1.8985065 -0.67076156  0.7024886 -0.5805955  0.717888413
#>             [,19]      [,20]
#>  [1,] -5.05364861 -1.9944701
#>  [2,]  2.34622934 -2.2950162
#>  [3,]  1.04992124  2.2080536
#>  [4,] -0.07920224  0.2728514
#>  [5,] -4.87435336 -2.6423441
#>  [6,]  0.11221185 -0.4242846
#>  [7,] -0.29874747  4.2554131
#>  [8,] -4.93667900 -0.5958184
#>  [9,] -5.18188550  3.2148704
#> [10,] -5.53891152  1.7433678
#> [11,] -0.01693155  0.7087705
#> [12,]  0.18231779  2.2885913
#> [13,]  3.79772080 -2.2337813
#> [14,] -1.90068229 -1.8985065
#> [15,]  2.54162888 -0.6707616
#> [16,]  3.25956870  0.7024886
#> [17,]  1.43925384 -0.5805955
#> [18,] -0.40981781  0.7178884
#> [19,] 11.78585618 -1.1267745
#> [20,] -1.12677454  4.4961227
#> 
#> , , 2
#> 
#>             [,1]        [,2]       [,3]         [,4]        [,5]        [,6]
#>  [1,]  3.8484893 -4.83075880  1.0737758 -0.203897970 -0.24204741  2.99070959
#>  [2,] -4.8307588 10.45951810 -0.1513977  0.068048259  1.80255819 -2.80946946
#>  [3,]  1.0737758 -0.15139768  4.1878242 -0.540593665 -0.71192073  1.54631938
#>  [4,] -0.2038980  0.06804826 -0.5405937  1.553755845  0.98294440 -1.17050904
#>  [5,] -0.2420474  1.80255819 -0.7119207  0.982944403  1.84193473 -0.06804196
#>  [6,]  2.9907096 -2.80946946  1.5463194 -1.170509038 -0.06804196  3.69040221
#>  [7,]  0.4985672 -1.95494292  2.0833371 -1.716436672 -1.90581181  1.47442515
#>  [8,]  0.1445636  1.20598930  0.3211217 -1.121936154  0.51808055  1.79859747
#>  [9,] -1.2022210  3.35604564  1.0684358  1.193458342  1.82826914 -0.59969552
#> [10,] -2.3894342  4.73864670 -1.3309241  0.108809389  0.62875617 -2.01599735
#> [11,] -0.6593582  0.75630680 -0.8394469  0.146597719  0.54789651 -0.27444819
#> [12,]  1.3937264 -2.64542325 -0.8848601 -0.945752235 -0.76497133  1.35206532
#> [13,] -1.4661485  1.05411069 -0.9000038 -0.002915795 -0.06233037 -1.11876602
#> [14,]  1.2661542 -0.52749358  0.4935821  1.035188264  2.33001123  1.93026295
#> [15,]  2.8097594 -5.05066925  3.5307165 -0.505326693 -1.94416627  1.96481085
#> [16,]  1.3624539 -3.38911503 -1.7323398 -0.068099820 -0.22356017  0.77287209
#> [17,] -1.7952391  1.39854011 -1.0585358  1.122676894 -0.57894389 -3.38734901
#> [18,]  0.1197260  0.72052208 -1.2594325  0.779872758  1.43292722  0.02054228
#> [19,] -0.9400735  4.58001063 -0.5768829  0.027651826  0.50644454 -1.30175167
#> [20,] -0.9762928  1.68770486 -3.4128207 -0.304670748  1.63193838  0.35710289
#>              [,7]         [,8]        [,9]       [,10]       [,11]       [,12]
#>  [1,]  0.49856719  0.144563589 -1.20222105 -2.38943415 -0.65935819  1.39372640
#>  [2,] -1.95494292  1.205989304  3.35604564  4.73864670  0.75630680 -2.64542325
#>  [3,]  2.08333711  0.321121671  1.06843583 -1.33092408 -0.83944690 -0.88486014
#>  [4,] -1.71643667 -1.121936154  1.19345834  0.10880939  0.14659772 -0.94575223
#>  [5,] -1.90581181  0.518080554  1.82826914  0.62875617  0.54789651 -0.76497133
#>  [6,]  1.47442515  1.798597470 -0.59969552 -2.01599735 -0.27444819  1.35206532
#>  [7,]  3.46506110  0.827797399 -1.23236664 -1.63050414 -0.39403256  0.79427385
#>  [8,]  0.82779740  2.232915596  0.44180015 -0.05377761  0.43421280  0.47780203
#>  [9,] -1.23236664  0.441800153  3.30619278  0.59843262  0.46184951 -2.04855963
#> [10,] -1.63050414 -0.053777609  0.59843262  2.89585542  0.29473588 -0.79575209
#> [11,] -0.39403256  0.434212796  0.46184951  0.29473588  0.46082836 -0.09319864
#> [12,]  0.79427385  0.477802030 -2.04855963 -0.79575209 -0.09319864  1.71215688
#> [13,]  0.05729291 -0.005682689  0.08416976  0.58717534  0.39797508 -0.19907734
#> [14,] -0.91434868  1.548542913  2.71641546 -1.54811296  0.75933558 -0.62914757
#> [15,]  2.80722199 -0.871866685 -1.12838344 -3.09618292 -1.23092920  0.46834975
#> [16,] -0.08336509 -0.075201059 -1.64659560 -1.06445001  0.16259428  1.49134177
#> [17,] -1.50100264 -2.634700043 -0.42962821  1.78643570 -0.37477295 -0.94354224
#> [18,] -1.76407231  0.315888159  0.89899492  0.45536467  0.46872623 -0.16603503
#> [19,] -2.29453092 -0.505051929  0.01077451  3.27025678 -0.43565742 -0.55451842
#> [20,] -1.34307483  2.073887153  0.18106255  1.07530458  1.37127548  0.92286964
#>              [,13]      [,14]       [,15]       [,16]      [,17]       [,18]
#>  [1,] -1.466148483  1.2661542  2.80975940  1.36245388 -1.7952391  0.11972598
#>  [2,]  1.054110686 -0.5274936 -5.05066925 -3.38911503  1.3985401  0.72052208
#>  [3,] -0.900003817  0.4935821  3.53071651 -1.73233985 -1.0585358 -1.25943245
#>  [4,] -0.002915795  1.0351883 -0.50532669 -0.06809982  1.1226769  0.77987276
#>  [5,] -0.062330366  2.3300112 -1.94416627 -0.22356017 -0.5789439  1.43292722
#>  [6,] -1.118766020  1.9302629  1.96481085  0.77287209 -3.3873490  0.02054228
#>  [7,]  0.057292912 -0.9143487  2.80722199 -0.08336509 -1.5010026 -1.76407231
#>  [8,] -0.005682689  1.5485429 -0.87186669 -0.07520106 -2.6347000  0.31588816
#>  [9,]  0.084169757  2.7164155 -1.12838344 -1.64659560 -0.4296282  0.89899492
#> [10,]  0.587175340 -1.5481130 -3.09618292 -1.06445001  1.7864357  0.45536467
#> [11,]  0.397975083  0.7593356 -1.23092920  0.16259428 -0.3747729  0.46872623
#> [12,] -0.199077340 -0.6291476  0.46834975  1.49134177 -0.9435422 -0.16603503
#> [13,]  0.768544581 -0.4175906 -1.11872491 -0.03350775  0.4750526 -0.06427155
#> [14,] -0.417590601  5.3654044 -0.45223584  0.22705954 -3.1997485  1.67867915
#> [15,] -1.118724909 -0.4522358  5.57403918  0.04348013 -0.6306952 -1.78315584
#> [16,] -0.033507750  0.2270595  0.04348013  1.96836125 -0.5932829  0.37867133
#> [17,]  0.475052585 -3.1997485 -0.63069517 -0.59328294  4.2236720 -0.39827021
#> [18,] -0.064271548  1.6786791 -1.78315584  0.37867133 -0.3982702  1.30955269
#> [19,] -0.438496823 -2.5495334 -2.27921178 -1.30890218  2.4122050  0.45428523
#> [20,]  0.859189757  1.9799478 -4.33176938  1.34702225 -1.9520064  1.74865590
#>             [,19]      [,20]
#>  [1,] -0.94007349 -0.9762928
#>  [2,]  4.58001063  1.6877049
#>  [3,] -0.57688290 -3.4128207
#>  [4,]  0.02765183 -0.3046707
#>  [5,]  0.50644454  1.6319384
#>  [6,] -1.30175167  0.3571029
#>  [7,] -2.29453092 -1.3430748
#>  [8,] -0.50505193  2.0738872
#>  [9,]  0.01077451  0.1810626
#> [10,]  3.27025678  1.0753046
#> [11,] -0.43565742  1.3712755
#> [12,] -0.55451842  0.9228696
#> [13,] -0.43849682  0.8591898
#> [14,] -2.54953337  1.9799478
#> [15,] -2.27921178 -4.3317694
#> [16,] -1.30890218  1.3470223
#> [17,]  2.41220504 -1.9520064
#> [18,]  0.45428523  1.7486559
#> [19,]  5.79144222 -0.3650459
#> [20,] -0.36504586  5.4368110
#>